refactor: 70.J.1.c — extract 8 Buffer impl groups into submodules
Complete the buffer.rs god-file split by extracting 8 logical groups of
Buffer methods into sibling submodules under freminal-buffer/src/buffer/:

  - cursor.rs        — cursor position/state operations
  - erase.rs         — ED/EL/DCH/ECH/ICH erase operations
  - flatten.rs       — row flattening, snapshot building, text extraction
  - images.rs        — image placement operations
  - lifecycle.rs     — construction, reset, invariants, prompt tracking
  - lines.rs         — LF/IND/NEL/RI/IL/DL/BS/DECALN
  - resize_and_alt.rs— resize, reflow, format/mode setters, alt-screen switch
  - scroll.rs        — scroll region, margins, scroll up/down/back/forward

buffer/mod.rs now contains:
  - Buffer and SavedPrimaryState struct definitions
  - Three tightly-coupled insertion methods that stay coupled with core
    wrap logic (insert_text, advance_row_for_wrap, reuse_row_as_softwrap)
  - Free helper functions (clamped_offset, text_col_span, tags_same_format)
  - All test modules (kept in place for now — moving tests is a separate
    concern and not required by the plan)

Before: buffer.rs = 11,024 lines (single file)
After:  buffer/mod.rs = 7,483 lines (~4,000 production + test modules
        unchanged) + 7 focused submodules totaling ~3,764 lines of
        production code clearly organized by responsibility.

Production Buffer impl code was reduced in mod.rs from ~4,000 lines to
~450 lines (the remaining insertion core + struct defs + free helpers).

No logic or signatures changed — pure move. Buffer struct fields
use pub(in crate::buffer) visibility so submodules can access them
without leaking internals to the rest of the crate.

Verification: cargo test --all green, cargo clippy --all-targets
--all-features -- -D warnings clean, cargo machete clean.
