+++ freminal-buffer/src/buffer/mod.rs
@@ -6,22 +6,34 @@
 use conv2::ValueFrom;
 use freminal_common::buffer_states::{
     buffer_type::BufferType,
-    cursor::{CursorPos, CursorState},
+    cursor::CursorState,
     format_tag::FormatTag,
-    modes::{
-        decawm::Decawm, declrmm::Declrmm, decom::Decom, lnm::Lnm,
-        reverse_wrap_around::ReverseWrapAround, xt_rev_wrap2::XtRevWrap2,
-    },
+    modes::{decawm::Decawm, declrmm::Declrmm, decom::Decom, lnm::Lnm},
     tchar::TChar,
 };
 
 use crate::{
-    cell::Cell,
-    image_store::{ImagePlacement, ImageProtocol, ImageStore, InlineImage},
+    image_store::ImageStore,
     response::InsertResponse,
     row::{Row, RowJoin, RowOrigin},
 };
 
+#[cfg(test)]
+use crate::cell::Cell;
+#[cfg(test)]
+use freminal_common::buffer_states::{
+    cursor::CursorPos,
+    modes::{reverse_wrap_around::ReverseWrapAround, xt_rev_wrap2::XtRevWrap2},
+};
+
+mod cursor;
+mod erase;
+mod flatten;
+mod images;
+mod lifecycle;
+mod lines;
+mod resize_and_alt;
+mod scroll;
 mod tabs;
 
 /// Apply a signed `delta` (in cells or rows) to an unsigned `base` coordinate and clamp the
@@ -207,53 +219,6 @@ fn text_col_span(text: &[TChar], col: usize, wrap_col: usize) -> usize {
 }
 
 impl Buffer {
-    /// Generate default tab stops at every 8 columns for the given width.
-    fn default_tab_stops(width: usize) -> Vec<bool> {
-        let mut stops = vec![false; width];
-        for i in (8..width).step_by(8) {
-            stops[i] = true;
-        }
-        stops
-    }
-
-    /// Creates a new Buffer with the specified width and height.
-    #[must_use]
-    pub fn new(width: usize, height: usize) -> Self {
-        // Start with a single blank row.  The buffer grows dynamically as
-        // content is written.  Pre-allocating `height` empty rows caused the
-        // visible area to always contain `height` rows, most of which were
-        // blank — the GUI's stick_to_bottom would then display those trailing
-        // blank rows instead of the actual content at the top.
-        let rows = vec![Row::new(width)];
-        let row_cache = vec![None];
-
-        Self {
-            rows,
-            row_cache,
-            width,
-            height,
-            cursor: CursorState::default(),
-            current_tag: FormatTag::default(),
-            scrollback_limit: 4000,
-            kind: BufferType::Primary,
-            saved_primary: None,
-            saved_cursor: None,
-            lnm_enabled: Lnm::LineFeed,
-            wrap_enabled: Decawm::AutoWrap,
-            preserve_scrollback_anchor: false,
-            scroll_region_top: 0,
-            scroll_region_bottom: height.saturating_sub(1),
-            scroll_region_left: 0,
-            scroll_region_right: width.saturating_sub(1),
-            declrmm_enabled: Declrmm::Disabled,
-            tab_stops: Self::default_tab_stops(width),
-            decom_enabled: Decom::NormalCursor,
-            image_store: ImageStore::new(),
-            image_cell_count: 0,
-            prompt_rows: Vec::new(),
-        }
-    }
-
     /// Return a new buffer with the given scrollback limit instead of the
     /// default (4000).  This is a builder-style method intended for
     /// production use where the value comes from user configuration.
@@ -263,184 +228,10 @@ impl Buffer {
         self
     }
 
-    /// Full terminal reset (RIS — Reset to Initial State).
-    ///
-    /// Restores the buffer to its initial startup state:
-    /// - Clears all screen content and scrollback
-    /// - Resets cursor to home position (0,0)
-    /// - Resets all character attributes
-    /// - Resets scroll region to full screen
-    /// - Resets tab stops to default 8-column positions
-    /// - Exits alternate buffer if active
-    ///
-    /// Preserves `width`, `height`, and `scrollback_limit` (terminal geometry
-    /// and user configuration).
-    pub fn full_reset(&mut self) {
-        self.rows = vec![Row::new(self.width)];
-        self.row_cache = vec![None];
-        self.cursor = CursorState::default();
-        self.current_tag = FormatTag::default();
-        self.kind = BufferType::Primary;
-        self.saved_primary = None;
-        self.saved_cursor = None;
-        self.lnm_enabled = Lnm::LineFeed;
-        self.wrap_enabled = Decawm::AutoWrap;
-        self.preserve_scrollback_anchor = false;
-        self.scroll_region_top = 0;
-        self.scroll_region_bottom = self.height.saturating_sub(1);
-        self.scroll_region_left = 0;
-        self.scroll_region_right = self.width.saturating_sub(1);
-        self.declrmm_enabled = Declrmm::Disabled;
-        self.tab_stops = Self::default_tab_stops(self.width);
-        self.decom_enabled = Decom::NormalCursor;
-        self.image_store.clear();
-        self.image_cell_count = 0;
-        self.prompt_rows.clear();
-    }
-
     /// The maximum number of off-screen rows retained above the visible area.
     #[must_use]
     pub const fn scrollback_limit(&self) -> usize {
         self.scrollback_limit
-    }
-
-    /// Record the current cursor row as a prompt-start marker.
-    ///
-    /// Called by `TerminalHandler` when an OSC 133 `PromptStart` fires.
-    pub fn mark_prompt_row(&mut self) {
-        self.prompt_rows.push(self.cursor.pos.y);
-    }
-
-    /// Buffer-relative row indices of all recorded prompt-start markers.
-    #[must_use]
-    pub fn prompt_rows(&self) -> &[usize] {
-        &self.prompt_rows
-    }
-
-    /// Shift all prompt-row markers down by `removed` and drop any that
-    /// fell below zero.  Called after draining rows from the front.
-    fn adjust_prompt_rows(&mut self, removed: usize) {
-        self.prompt_rows.retain_mut(|r| {
-            r.checked_sub(removed).is_some_and(|adjusted| {
-                *r = adjusted;
-                true
-            })
-        });
-    }
-
-    /// Internal consistency checks for debug builds.
-    ///
-    /// This is called from most mutating entry points. In release builds
-    /// it compiles down to a no-op.
-    #[cfg(debug_assertions)]
-    fn debug_assert_invariants(&self) {
-        // If there are no rows at all, we expect a fully reset buffer state.
-        if self.rows.is_empty() {
-            debug_assert_eq!(self.cursor.pos.y, 0, "empty buffer must keep cursor.y at 0");
-            debug_assert_eq!(self.cursor.pos.x, 0, "empty buffer must keep cursor.x at 0");
-            return;
-        }
-
-        // Cursor Y must always point at an existing row.
-        debug_assert!(
-            self.cursor.pos.y < self.rows.len(),
-            "cursor.pos.y {} out of bounds for rows.len() {}",
-            self.cursor.pos.y,
-            self.rows.len()
-        );
-
-        // Cursor X must be within [0, width) if width > 0.
-        if self.width == 0 {
-            debug_assert_eq!(
-                self.cursor.pos.x, 0,
-                "width=0 buffer must keep cursor.x at 0"
-            );
-        } else {
-            debug_assert!(
-                self.cursor.pos.x <= self.width,
-                "cursor.pos.x {} out of bounds for width {}",
-                self.cursor.pos.x,
-                self.width
-            );
-        }
-
-        // Scrollback invariants by buffer kind.
-        match self.kind {
-            BufferType::Primary => {
-                // Primary buffer: rows must never exceed height + scrollback_limit.
-                let max_rows = self.height + self.scrollback_limit;
-                debug_assert!(
-                    self.rows.len() <= max_rows,
-                    "primary buffer has {} rows but max_rows is {} (height={} + scrollback_limit={})",
-                    self.rows.len(),
-                    max_rows,
-                    self.height,
-                    self.scrollback_limit
-                );
-            }
-            BufferType::Alternate => {
-                // Alternate buffer: fixed-size, no scrollback.
-                debug_assert_eq!(
-                    self.rows.len(),
-                    self.height,
-                    "alternate buffer must have exactly `height` rows (got rows.len()={}, height={})",
-                    self.rows.len(),
-                    self.height
-                );
-            }
-        }
-
-        // Scroll region (DECSTBM) invariants: screen-relative.
-        if self.height > 0 {
-            debug_assert!(
-                self.scroll_region_top <= self.scroll_region_bottom,
-                "scroll_region_top {} must be <= scroll_region_bottom {}",
-                self.scroll_region_top,
-                self.scroll_region_bottom
-            );
-            debug_assert!(
-                self.scroll_region_bottom < self.height,
-                "scroll_region_bottom {} must be < height {}",
-                self.scroll_region_bottom,
-                self.height
-            );
-        }
-
-        // Cache length must always match rows length.
-        debug_assert_eq!(
-            self.row_cache.len(),
-            self.rows.len(),
-            "row_cache length {} != rows length {}",
-            self.row_cache.len(),
-            self.rows.len()
-        );
-
-        // Image cell count must match the actual number of image cells across
-        // all rows.  This is O(rows × cols) but only runs in debug builds.
-        let actual_image_cells: usize = self.rows.iter().map(Row::count_image_cells).sum();
-        debug_assert_eq!(
-            self.image_cell_count, actual_image_cells,
-            "image_cell_count {} != actual image cells {}",
-            self.image_cell_count, actual_image_cells
-        );
-    }
-
-    // In release builds this is a no-op, so we can call it freely.
-    #[cfg(not(debug_assertions))]
-    #[inline]
-    fn debug_assert_invariants(&self) {}
-
-    fn push_row(&mut self, origin: RowOrigin, join: RowJoin) {
-        let row = Row::new_with_origin(self.width, origin, join);
-        // New rows created by scrolling (LF at bottom, auto-wrap at bottom-right)
-        // use default background — NOT the current SGR background.  BCE
-        // (back_color_erase) only applies to explicit erase operations (ED, EL).
-        // Filling with current_tag here causes visible artifacts when programs
-        // output long lines with colored backgrounds that wrap at the right margin:
-        // the trailing blank cells on the wrapped continuation row retain the
-        // non-default background instead of being transparent.
-        self.rows.push(row);
-        self.row_cache.push(None);
     }
 
     /// Returns a reference to all rows in this buffer (scrollback + visible region).
@@ -453,135 +244,6 @@ impl Buffer {
     #[must_use]
     pub const fn get_cursor(&self) -> &CursorState {
         &self.cursor
-    }
-
-    /// Set the cursor to an absolute buffer position without any DECOM or
-    /// screen-relative translation.
-    ///
-    /// The position is clamped to the current buffer dimensions.  Used by
-    /// DECSDM to restore the cursor after `place_image` moves it.
-    pub fn set_cursor_pos_raw(&mut self, pos: CursorPos) {
-        self.cursor.pos.x = if self.width > 0 {
-            pos.x.min(self.width - 1)
-        } else {
-            0
-        };
-        self.cursor.pos.y = pos.y.min(self.rows.len().saturating_sub(1));
-    }
-
-    /// Advance the cursor by one column, wrapping to the next line if needed.
-    ///
-    /// Used after inserting a placeholder image cell that occupies one column.
-    pub const fn advance_cursor_one(&mut self) {
-        if self.cursor.pos.x + 1 < self.width {
-            self.cursor.pos.x += 1;
-        }
-        // If at the rightmost column, don't wrap automatically — let the
-        // next character insertion handle wrap/scroll as normal.
-    }
-
-    /// Set the [`LineWidth`] attribute on the row under the cursor.
-    ///
-    /// This is the buffer-level primitive called by the terminal handler in
-    /// response to `ESC # 3` (double-height top), `ESC # 4` (double-height
-    /// bottom), `ESC # 5` (single-width), and `ESC # 6` (double-width).
-    ///
-    /// The row is marked dirty so the next snapshot rebuild re-flattens it.
-    pub fn set_cursor_line_width(&mut self, lw: crate::row::LineWidth) {
-        let row_idx = self.cursor.pos.y;
-        if let Some(row) = self.rows.get_mut(row_idx)
-            && row.line_width != lw
-        {
-            row.line_width = lw;
-            row.dirty = true;
-        }
-    }
-
-    /// Return the [`LineWidth`] for each row in the visible window.
-    ///
-    /// The returned vector has `min(term_height, row_count)` entries, one per
-    /// visible row in top-to-bottom order.  Used by `build_snapshot` to thread
-    /// per-row line-width data through to the renderer.
-    #[must_use]
-    pub fn visible_line_widths(&self, scroll_offset: usize) -> Vec<crate::row::LineWidth> {
-        let start = self.visible_window_start(scroll_offset);
-        let end = (start + self.height).min(self.rows.len());
-        self.rows[start..end].iter().map(|r| r.line_width).collect()
-    }
-
-    /// Set an image cell at a specific (row, col) position in the buffer.
-    ///
-    /// Also invalidates the corresponding row cache entry.  Used by
-    /// `TerminalHandler` for Kitty Unicode placeholder cells.
-    pub fn set_image_cell_at(
-        &mut self,
-        row_idx: usize,
-        col_idx: usize,
-        placement: ImagePlacement,
-        tag: FormatTag,
-    ) {
-        if row_idx < self.rows.len() {
-            // Check if the old cell already had an image (avoid double-counting).
-            let had_image = self.rows[row_idx]
-                .cells()
-                .get(col_idx)
-                .is_some_and(Cell::has_image);
-            self.rows[row_idx].set_image_cell(col_idx, placement, tag);
-            if !had_image {
-                self.image_cell_count += 1;
-            }
-            if row_idx < self.row_cache.len() {
-                self.row_cache[row_idx] = None;
-            }
-        }
-    }
-
-    /// Return the cursor position in **screen coordinates** (0-indexed, relative
-    /// to the top of the visible window).
-    ///
-    /// Unlike `get_cursor().pos.y`, which is an absolute index into `self.rows`,
-    /// this subtracts `visible_window_start()` so the result is always in the
-    /// range `0..height` and matches what the GUI painter expects.
-    #[must_use]
-    pub fn get_cursor_screen_pos(&self) -> CursorPos {
-        let screen_y = self.cursor_screen_y();
-        CursorPos {
-            x: self.cursor.pos.x,
-            y: screen_y,
-        }
-    }
-
-    /// Get the rows that should be *visually displayed* in the GUI.
-    ///
-    /// Contract:
-    /// - Returns a contiguous slice of `self.rows`.
-    /// - `visible_rows(scroll_offset).len() <= self.height`.
-    /// - When `self.rows.len() <= self.height`, the slice is the entire buffer.
-    /// - When `scroll_offset == 0`, the slice is the last `height` rows
-    ///   (the live bottom).
-    /// - When `scroll_offset > 0`, the slice is shifted upwards into
-    ///   scrollback, clamped so it never goes before the oldest row.
-    /// - Never allocates; always borrows from `self.rows`.
-    ///
-    /// `scroll_offset` is owned by the caller (e.g. `ViewState`) and is never
-    /// stored inside `Buffer`.
-    #[must_use]
-    pub fn visible_rows(&self, scroll_offset: usize) -> &[Row] {
-        if self.rows.is_empty() {
-            return &[];
-        }
-
-        let total = self.rows.len();
-        let h = self.height;
-
-        // Clamp scroll_offset within bounds.
-        let max_offset = self.max_scroll_offset();
-        let offset = scroll_offset.min(max_offset);
-
-        let start = total.saturating_sub(h + offset);
-        let end = start + h;
-
-        &self.rows[start.min(total)..end.min(total)]
     }
 
     /// Reset an existing row at `row_idx` to a soft-wrap continuation and update
@@ -793,3143 +455,6 @@ impl Buffer {
         self.cursor.pos.y = new_row_idx;
         new_row_idx
     }
-
-    /// Resize the terminal buffer and return the adjusted `scroll_offset`.
-    ///
-    /// The caller passes in the current `scroll_offset` (from `ViewState`) and
-    /// receives back the clamped/reset value that should be stored into `ViewState`.
-    pub fn set_size(&mut self, new_width: usize, new_height: usize, scroll_offset: usize) -> usize {
-        let width_changed = new_width != self.width;
-        let height_changed = new_height != self.height;
-
-        if !width_changed && !height_changed {
-            return scroll_offset;
-        }
-
-        // ---- WIDTH CHANGE → REFLOW ----
-        // Alternate buffers must never reflow (they represent a fixed-size screen,
-        // not a scrollback history).  Reflow re-wraps logical lines which can
-        // create more or fewer rows than `height`, breaking the invariant that
-        // alternate buffers always have exactly `height` rows.  Instead, just
-        // update each row's max_width (content that extends beyond the new width
-        // is simply clipped, which matches xterm/VT behaviour).
-        let after_reflow = if width_changed && self.kind != BufferType::Alternate {
-            self.reflow_to_width(new_width);
-            0
-        } else {
-            scroll_offset
-        };
-
-        // ---- HEIGHT CHANGE → GROW/SHRINK SCREEN ----
-        // For alternate buffers we must ALWAYS reconcile row count with the
-        // target height, even when only the width changed (reflow was skipped
-        // above, so rows.len() is still the old height — but we also guard
-        // against any future code path that could desync them).
-        let needs_height_adjust =
-            height_changed || (self.kind == BufferType::Alternate && self.rows.len() != new_height);
-
-        let after_resize = if needs_height_adjust {
-            let adjusted = self.resize_height(new_height, after_reflow);
-
-            // Validate scroll region against new height.
-            //
-            // `self.height` is still the OLD height here (updated at line 785).
-            //
-            // If the scroll region covered the entire old screen (which is the
-            // default after DECSTBM reset or after `enter_alternate`), expand it
-            // to cover the entire NEW screen.  Without this, growing the buffer
-            // (e.g. from 29→58 rows when a split pane is closed) leaves the
-            // scroll region at the old height.  Full-screen TUIs like nvim that
-            // redraw via space-fill + LF rather than ED hit the old
-            // scroll_region_bottom and start scrolling prematurely, causing the
-            // lower half of the screen to never be written.
-            let old_max_bottom = self.height.saturating_sub(1);
-            let new_max_bottom = new_height.saturating_sub(1);
-
-            let was_full_screen =
-                self.scroll_region_top == 0 && self.scroll_region_bottom == old_max_bottom;
-
-            if was_full_screen {
-                // Region was full-screen → keep it full-screen at the new size.
-                self.scroll_region_top = 0;
-                self.scroll_region_bottom = new_max_bottom;
-            } else if self.scroll_region_bottom >= new_height
-                || self.scroll_region_top >= new_height
-                || self.scroll_region_top >= self.scroll_region_bottom
-            {
-                // Region is now invalid → reset to full screen.
-                self.scroll_region_top = 0;
-                self.scroll_region_bottom = new_max_bottom;
-            } else {
-                // Partial region still valid → just clamp bottom.
-                self.scroll_region_bottom = self.scroll_region_bottom.min(new_max_bottom);
-            }
-
-            adjusted
-        } else {
-            after_reflow
-        };
-
-        // Update buffer scalars
-        self.width = new_width;
-        self.height = new_height;
-
-        // Preserve existing tab stops across width changes:
-        // - Wider: extend with defaults (every 8th column) for new columns only.
-        // - Narrower: truncate to the new width.
-        if width_changed {
-            let old_width = self.tab_stops.len();
-            if new_width > old_width {
-                // Extend the vector with `false` for the new columns
-                self.tab_stops.resize(new_width, false);
-                // Set default 8-column stops only in the newly added range.
-                // Default stops are at 8, 16, 24, … (matching `default_tab_stops`).
-                for col in (8..new_width).step_by(8).filter(|&c| c >= old_width) {
-                    self.tab_stops[col] = true;
-                }
-            } else if new_width < old_width {
-                self.tab_stops.truncate(new_width);
-            }
-        }
-
-        // Ensure every row's max_width matches the new buffer width.
-        // For alternate buffers (which skip reflow), the row-level cache
-        // entries still contain flattened data at the old width — we must
-        // invalidate them so the next `flatten_visible` re-flattens every
-        // row at the new width.  Without this, nvim (and any full-screen
-        // TUI on the alternate screen) renders with stale row data after a
-        // resize, causing gaps, uncolored cells, and mispositioned content.
-        if width_changed {
-            for (i, row) in self.rows.iter_mut().enumerate() {
-                row.set_max_width(new_width);
-                // On the alternate screen (which skips reflow) a shrink leaves
-                // stale cells beyond the new width in row.cells.  flatten_row
-                // iterates row.cells directly, so those stale cells would leak
-                // into the snapshot and render as a strip of old content at
-                // the right edge of the viewport.  `truncate_cells_to_width`
-                // is a no-op when cells.len() <= new_width, so it is safe to
-                // call unconditionally (including on grow).
-                row.truncate_cells_to_width(new_width);
-                row.dirty = true;
-                self.row_cache[i] = None;
-            }
-        }
-
-        // Always clamp cursor after size change
-        self.clamp_cursor_after_resize();
-
-        // Enforce scrollback limit after resize (reflow may have created extra rows)
-        let final_offset = self.enforce_scrollback_limit(after_resize);
-
-        // When on the alternate screen, also resize the saved primary buffer
-        // so that `leave_alternate` restores a buffer that matches the current
-        // terminal dimensions.  Without this, exiting a full-screen TUI (e.g.
-        // nvim) after a pane resize restores the primary buffer at the old
-        // dimensions, causing immediate rendering artifacts.
-        if self.kind == BufferType::Alternate
-            && let Some(saved) = self.saved_primary.take()
-        {
-            let saved = Self::resize_saved_primary(saved, new_width, new_height);
-            self.saved_primary = Some(saved);
-        }
-
-        self.debug_assert_invariants();
-
-        final_offset
-    }
-
-    /// Resize a saved primary buffer to new dimensions.
-    ///
-    /// Builds a temporary primary `Buffer`, applies `set_size`, and extracts
-    /// the updated state back into a `SavedPrimaryState`.  This reuses all
-    /// the existing resize logic (reflow, height adjust, scroll region
-    /// validation, cursor clamping, scrollback limit enforcement) instead of
-    /// duplicating it.
-    fn resize_saved_primary(
-        saved: SavedPrimaryState,
-        new_width: usize,
-        new_height: usize,
-    ) -> SavedPrimaryState {
-        // Reconstruct a temporary primary Buffer from the saved state.
-        let old_width = saved.rows.first().map_or(new_width, Row::max_width);
-        let old_height = saved.height;
-
-        let mut tmp = Self {
-            rows: saved.rows,
-            row_cache: saved.row_cache,
-            width: old_width,
-            height: old_height,
-            cursor: saved.cursor,
-            current_tag: FormatTag::default(),
-            scrollback_limit: 4000,
-            kind: BufferType::Primary,
-            saved_primary: None,
-            saved_cursor: saved.saved_cursor,
-            lnm_enabled: Lnm::LineFeed,
-            wrap_enabled: Decawm::AutoWrap,
-            preserve_scrollback_anchor: false,
-            scroll_region_top: saved.scroll_region_top,
-            scroll_region_bottom: saved.scroll_region_bottom,
-            scroll_region_left: saved.scroll_region_left,
-            scroll_region_right: saved.scroll_region_right,
-            declrmm_enabled: Declrmm::Disabled,
-            tab_stops: Self::default_tab_stops(old_width),
-            decom_enabled: Decom::NormalCursor,
-            image_store: saved.image_store,
-            image_cell_count: saved.image_cell_count,
-            prompt_rows: Vec::new(),
-        };
-
-        let new_offset = tmp.set_size(new_width, new_height, saved.scroll_offset);
-
-        SavedPrimaryState {
-            rows: tmp.rows,
-            row_cache: tmp.row_cache,
-            cursor: tmp.cursor,
-            scroll_offset: new_offset,
-            height: new_height,
-            scroll_region_top: tmp.scroll_region_top,
-            scroll_region_bottom: tmp.scroll_region_bottom,
-            scroll_region_left: tmp.scroll_region_left,
-            scroll_region_right: tmp.scroll_region_right,
-            saved_cursor: tmp.saved_cursor,
-            image_store: tmp.image_store,
-            image_cell_count: tmp.image_cell_count,
-        }
-    }
-
-    // Inherently large: the reflow algorithm walks every logical line, splits/joins rows at the
-    // new width, and preserves all cell content and tags. The size reflects algorithmic
-    // complexity, not lack of structure.
-    #[allow(clippy::too_many_lines)]
-    /// Re-wrap all rows to `new_width` columns without losing any text.
-    ///
-    /// ## Algorithm
-    ///
-    /// 1. **Group into logical lines.** Rows are joined by a `RowJoin` flag:
-    ///    `NewLogicalLine` starts a new logical line and `ContinueLogicalLine`
-    ///    means this row is a soft-wrapped continuation of the previous one.
-    ///    We split `self.rows` into groups where each group represents one
-    ///    original logical line (a paragraph, in word-processor terms).
-    ///
-    /// 2. **Flatten each logical line.** The cells from every physical row in
-    ///    the group are concatenated into a single `Vec<Cell>`, discarding the
-    ///    old row boundaries.
-    ///
-    /// 3. **Re-wrap at the new width.** Wide glyphs (display width > 1) are
-    ///    kept whole — if a glyph does not fit at the current column it is
-    ///    moved to the next row rather than split.  Each flushed row records
-    ///    whether it is the first physical row of its logical line
-    ///    (`RowJoin::NewLogicalLine`) or a continuation (`ContinueLogicalLine`),
-    ///    and inherits the original `RowOrigin` of its logical line.
-    ///
-    /// 4. **Install the new rows.** `self.rows` is replaced with the reflow
-    ///    result, `self.width` is updated, and `self.row_cache` is reset to
-    ///    all-`None` (every row is dirty after reflow).
-    ///
-    /// The operation is O(total cells) — linear in the amount of text.
-    pub fn reflow_to_width(&mut self, new_width: usize) {
-        if new_width == 0 || self.rows.is_empty() || new_width == self.width {
-            // Nothing to do
-            return;
-        }
-
-        let old_cursor_y = self.cursor.pos.y;
-        let old_cursor_x = self.cursor.pos.x;
-
-        // Take ownership of the old rows
-        let old_rows = std::mem::take(&mut self.rows);
-
-        // 1) Group rows into logical lines based on RowJoin.
-        //    While grouping, identify which logical line contains the cursor
-        //    and compute the cursor's flat cell offset within that line.
-        let mut logical_lines: Vec<Vec<Row>> = Vec::new();
-        let mut current_line: Vec<Row> = Vec::new();
-
-        let mut cursor_logical_line: Option<usize> = None;
-        let mut cursor_flat_offset: usize = 0;
-
-        for (old_row_idx, row) in old_rows.into_iter().enumerate() {
-            if row.join == RowJoin::NewLogicalLine && !current_line.is_empty() {
-                logical_lines.push(current_line);
-                current_line = Vec::new();
-            }
-
-            if old_row_idx == old_cursor_y {
-                cursor_logical_line = Some(logical_lines.len());
-                // Flat offset = cells from preceding rows in this logical line + cursor X.
-                cursor_flat_offset = current_line
-                    .iter()
-                    .map(|r| r.get_characters().len())
-                    .sum::<usize>()
-                    + old_cursor_x;
-            }
-
-            current_line.push(row);
-        }
-        if !current_line.is_empty() {
-            logical_lines.push(current_line);
-        }
-
-        // 2) For each logical line, flatten its cells and re-wrap
-        let mut new_rows: Vec<Row> = Vec::new();
-        let mut new_cursor_y: Option<usize> = None;
-        let mut new_cursor_x: Option<usize> = None;
-
-        for (line_idx, line) in logical_lines.into_iter().enumerate() {
-            // Determine origin for the first row of this logical line.
-            let first_origin = line.first().map_or(RowOrigin::HardBreak, |r| r.origin);
-            let is_cursor_line = cursor_logical_line == Some(line_idx);
-
-            // Flatten all rows in this logical line into a single Vec<Cell>
-            let mut flat_cells: Vec<Cell> = Vec::new();
-            for row in &line {
-                flat_cells.extend(row.get_characters().iter().cloned());
-            }
-
-            // Record where this logical line's new rows start (for cursor mapping).
-            let line_start_idx = new_rows.len();
-
-            if flat_cells.is_empty() {
-                // Empty logical line → keep a single empty row
-                new_rows.push(Row::new_with_origin(
-                    new_width,
-                    first_origin,
-                    RowJoin::NewLogicalLine,
-                ));
-                if is_cursor_line {
-                    new_cursor_y = Some(new_rows.len() - 1);
-                    new_cursor_x = Some(old_cursor_x.min(new_width.saturating_sub(1)));
-                }
-                continue;
-            }
-
-            let mut idx = 0;
-            let mut col = 0;
-            let mut cur_cells: Vec<Cell> = Vec::new();
-            let mut is_first_row_for_line = true;
-
-            while idx < flat_cells.len() {
-                let cell = &flat_cells[idx];
-
-                if cell.is_head() {
-                    let w = cell.display_width().max(1);
-
-                    // If this glyph doesn't fit on the current row (and we already have content),
-                    // flush the current row and start a new one.
-                    if col + w > new_width && col > 0 {
-                        let origin = if is_first_row_for_line {
-                            first_origin
-                        } else {
-                            RowOrigin::SoftWrap
-                        };
-                        let join = if is_first_row_for_line {
-                            RowJoin::NewLogicalLine
-                        } else {
-                            RowJoin::ContinueLogicalLine
-                        };
-
-                        new_rows.push(Row::from_cells(new_width, origin, join, cur_cells));
-
-                        cur_cells = Vec::new();
-                        col = 0;
-                        is_first_row_for_line = false;
-                    }
-
-                    // Now place this glyph (head + continuations) onto the row.
-                    cur_cells.push(cell.clone());
-                    idx += 1;
-
-                    let mut consumed = 1;
-                    while consumed < w
-                        && idx < flat_cells.len()
-                        && flat_cells[idx].is_continuation()
-                    {
-                        cur_cells.push(flat_cells[idx].clone());
-                        idx += 1;
-                        consumed += 1;
-                    }
-
-                    col += w.min(new_width);
-                } else {
-                    // Stray continuation (should be rare): treat as width 1 column.
-                    if col + 1 > new_width && col > 0 {
-                        let origin = if is_first_row_for_line {
-                            first_origin
-                        } else {
-                            RowOrigin::SoftWrap
-                        };
-                        let join = if is_first_row_for_line {
-                            RowJoin::NewLogicalLine
-                        } else {
-                            RowJoin::ContinueLogicalLine
-                        };
-
-                        new_rows.push(Row::from_cells(new_width, origin, join, cur_cells));
-
-                        cur_cells = Vec::new();
-                        col = 0;
-                        is_first_row_for_line = false;
-                    }
-
-                    cur_cells.push(cell.clone());
-                    idx += 1;
-                    col += 1;
-                }
-            }
-
-            // Flush any remaining cells as the final row of this logical line.
-            if !cur_cells.is_empty() {
-                let origin = if is_first_row_for_line {
-                    first_origin
-                } else {
-                    RowOrigin::SoftWrap
-                };
-                let join = if is_first_row_for_line {
-                    RowJoin::NewLogicalLine
-                } else {
-                    RowJoin::ContinueLogicalLine
-                };
-
-                new_rows.push(Row::from_cells(new_width, origin, join, cur_cells));
-            }
-
-            // If this logical line contains the cursor, map the old cursor
-            // position to the correct new row and column.
-            if is_cursor_line {
-                let mut flat_col: usize = 0;
-                let mut found = false;
-                for (i, new_row) in new_rows[line_start_idx..].iter().enumerate() {
-                    let row_cells = new_row.get_characters().len();
-                    if flat_col + row_cells > cursor_flat_offset {
-                        new_cursor_y = Some(line_start_idx + i);
-                        new_cursor_x = Some(cursor_flat_offset - flat_col);
-                        found = true;
-                        break;
-                    }
-                    flat_col += row_cells;
-                }
-                if !found {
-                    // Cursor is past the end of content (in blank space).
-                    // Place it on the last row of this logical line.
-                    let last_row_idx = new_rows.len() - 1;
-                    let last_row_len = new_rows[last_row_idx].get_characters().len();
-                    new_cursor_y = Some(last_row_idx);
-                    new_cursor_x = Some(cursor_flat_offset.saturating_sub(flat_col) + last_row_len);
-                }
-            }
-        }
-
-        // 3) Install the new rows and update width
-        self.rows = new_rows;
-        // All rows are freshly constructed (dirty=true by construction), so
-        // the entire cache is invalid.  Reset it to match the new row count.
-        self.row_cache = vec![None; self.rows.len()];
-        self.width = new_width;
-        // Reflow rebuilds all rows from scratch; recount image cells so the
-        // counter stays accurate regardless of how reflow may have clipped or
-        // merged cells.
-        self.image_cell_count = self.rows.iter().map(Row::count_image_cells).sum();
-
-        // 4) Remap cursor position based on reflow tracking.
-        if let (Some(cy), Some(cx)) = (new_cursor_y, new_cursor_x) {
-            self.cursor.pos.y = cy.min(self.rows.len().saturating_sub(1));
-            self.cursor.pos.x = cx.min(self.width.saturating_sub(1));
-        } else if self.cursor.pos.y >= self.rows.len() {
-            if self.rows.is_empty() {
-                self.cursor.pos.y = 0;
-            } else {
-                self.cursor.pos.y = self.rows.len() - 1;
-            }
-            self.cursor.pos.x = 0;
-        } else if self.cursor.pos.x >= self.width {
-            self.cursor.pos.x = self.width.saturating_sub(1);
-        }
-    }
-
-    /// Adjust the buffer rows for a new height and return the adjusted `scroll_offset`.
-    fn resize_height(&mut self, new_height: usize, scroll_offset: usize) -> usize {
-        let old_height = self.height;
-
-        if new_height > old_height {
-            // Grow: add blank rows at the bottom.
-            let grow = new_height - old_height;
-            for _ in 0..grow {
-                self.rows.push(Row::new(self.width));
-                self.row_cache.push(None);
-            }
-            // Mark all pre-existing rows dirty so the row-level cache is
-            // invalidated.  The visible window is now taller and the old
-            // snapshot was built for a different height — returning stale
-            // cached rows causes gaps and mispositioned content in
-            // full-screen TUIs (e.g. nvim) that rely on absolute cursor
-            // positioning after SIGWINCH.
-            for i in 0..old_height.min(self.rows.len()) {
-                self.rows[i].dirty = true;
-                self.row_cache[i] = None;
-            }
-        } else if new_height < old_height {
-            if self.kind == BufferType::Alternate {
-                // Alternate buffer must never have scrollback.  When shrinking,
-                // keep the bottom `new_height` rows (the ones the user can see)
-                // and discard the top ones.  This preserves the invariant
-                // `rows.len() == height` which all alternate-buffer coordinate
-                // logic depends on (handle_lf, handle_ri, insert_lines,
-                // delete_lines all compare cursor.pos.y against
-                // scroll_region_top/bottom directly without an offset).
-                let excess = self.rows.len().saturating_sub(new_height);
-                if excess > 0 {
-                    // Account for image cells in the drained rows.
-                    if self.image_cell_count > 0 {
-                        let drained_images: usize =
-                            self.rows[..excess].iter().map(Row::count_image_cells).sum();
-                        self.image_cell_count -= drained_images;
-                    }
-                    self.rows.drain(0..excess);
-                    self.row_cache.drain(0..excess);
-                    self.adjust_prompt_rows(excess);
-                    // Adjust cursor Y for the removed rows.
-                    self.cursor.pos.y = self.cursor.pos.y.saturating_sub(excess);
-                }
-            } else {
-                // Primary buffer: extra rows become scrollback (handled by
-                // enforce_scrollback_limit later).  The cursor's absolute row
-                // index is left unchanged — `clamp_cursor_after_resize()` (below)
-                // ensures it stays within `rows.len()`, and `visible_window_start(0)`
-                // already anchors the visible window to the bottom of the buffer.
-                //
-                // Previous code clamped `cursor.pos.y` to `new_height - 1`, which
-                // was wrong: `cursor.pos.y` is an absolute index into `self.rows`,
-                // not a screen-relative position.  That clamp moved the cursor from
-                // its real row (e.g. the shell prompt) into the middle of whatever
-                // content happened to be at row `new_height - 1`, causing the
-                // SIGWINCH-triggered shell redraw to overwrite that content.
-            }
-        }
-
-        if self.preserve_scrollback_anchor {
-            // IMPORTANT: use new_height, not self.height (which is still old here)
-            let max_offset = if self.rows.len() > new_height {
-                self.rows.len() - new_height
-            } else {
-                0
-            };
-            scroll_offset.min(max_offset)
-        } else {
-            // xterm-style: reset to live bottom
-            0
-        }
-    }
-
-    const fn clamp_cursor_after_resize(&mut self) {
-        // Clamp Y
-        if self.cursor.pos.y >= self.rows.len() {
-            self.cursor.pos.y = self.rows.len().saturating_sub(1);
-        }
-
-        // Clamp X
-        if self.cursor.pos.x >= self.width {
-            self.cursor.pos.x = self.width.saturating_sub(1);
-        }
-
-        if self.rows.is_empty() {
-            self.cursor.pos.x = 0;
-            self.cursor.pos.y = 0;
-        }
-    }
-
-    /// Enforce the scrollback row limit, adjusting the caller's `scroll_offset` if
-    /// rows are trimmed from the top.  Returns the (possibly reduced) scroll offset
-    /// that the caller should store into `ViewState`.
-    #[must_use]
-    fn enforce_scrollback_limit(&mut self, scroll_offset: usize) -> usize {
-        // Only primary buffer keeps scrollback.
-        if self.kind == BufferType::Alternate {
-            return scroll_offset;
-        }
-
-        let max_rows = self.height + self.scrollback_limit;
-
-        // Nothing to trim, but still make sure scroll_offset is not insane.
-        if self.rows.len() <= max_rows {
-            let max_offset = self.max_scroll_offset();
-            return scroll_offset.min(max_offset);
-        }
-
-        // Number of rows to drop from the top of the scrollback.
-        let overflow = self.rows.len() - max_rows;
-
-        // --- Adjust scroll_offset BEFORE modifying the rows ---
-        //
-        // If the user is scrolled back into the area we're about to delete,
-        // reduce their offset by the number of deleted rows. If that wipes
-        // out all their scrollback, snap them to live view.
-        let adjusted_offset = scroll_offset.saturating_sub(overflow);
-
-        // --- Drop the oldest rows (and their cache entries) ---
-        // First, account for any image cells in the rows being drained.
-        if self.image_cell_count > 0 {
-            let drained_images: usize = self.rows[..overflow]
-                .iter()
-                .map(Row::count_image_cells)
-                .sum();
-            self.image_cell_count -= drained_images;
-        }
-        self.rows.drain(0..overflow);
-        self.row_cache.drain(0..overflow);
-        self.adjust_prompt_rows(overflow);
-
-        // --- Garbage-collect images no longer referenced by any row ---
-        if !self.image_store.is_empty() {
-            self.image_store
-                .retain_referenced(self.rows.iter().map(Row::cells));
-        }
-
-        // --- Adjust cursor row index ---
-        //
-        // Cursor is always measured relative to self.rows, so subtract
-        // `overflow` from its y-coordinate when possible.
-        if self.cursor.pos.y >= overflow {
-            self.cursor.pos.y -= overflow;
-        } else {
-            self.cursor.pos.y = 0;
-        }
-
-        // Finally, clamp scroll_offset to the new max_scroll_offset().
-        let max_offset = self.max_scroll_offset();
-        let final_offset = adjusted_offset.min(max_offset);
-
-        self.debug_assert_invariants();
-
-        final_offset
-    }
-
-    /// Handle ANSI Backspace (BS, 0x08).
-    ///
-    /// Semantics (ECMA-48, VT100):
-    /// - Move cursor left by one cell.
-    /// - If the cursor is in the pending-wrap state (`cursor.pos.x == self.width`,
-    ///   i.e. just wrote into the last column), BS treats the cursor as if it were
-    ///   at `self.width - 1` (the last visible column) before subtracting, which
-    ///   places it at `self.width - 2`.  This matches real VT100 hardware, where
-    ///   the pending-wrap flag is separate from the reported column position and BS
-    ///   clears the flag and moves left from the reported column.
-    /// - If the cell to the left is a continuation cell of a wide glyph,
-    ///   skip left until the glyph head.
-    /// - If cursor is at column 0 and `reverse_wrap` is false, do nothing.
-    /// - If cursor is at column 0 and `reverse_wrap` is true, wrap to the
-    ///   last column of the previous line (within the visible screen, or
-    ///   into scrollback if `xt_rev_wrap2` is also true).
-    /// - Never deletes characters.
-    pub fn handle_backspace(&mut self, reverse_wrap: ReverseWrapAround, xt_rev_wrap2: XtRevWrap2) {
-        if self.cursor.pos.x == 0 {
-            if reverse_wrap == ReverseWrapAround::DontWrap {
-                return;
-            }
-            self.reverse_wrap_up(xt_rev_wrap2);
-            return;
-        }
-
-        let row_idx = self.cursor.pos.y;
-
-        if row_idx >= self.rows.len() {
-            return;
-        }
-
-        let row = &self.rows[row_idx];
-
-        // When in pending-wrap state the internal x equals self.width (one past
-        // the last column).  A real VT100 keeps the cursor at the last column with
-        // a separate pending-wrap bit; BS clears the bit and moves left from the
-        // reported column, landing at width-2 (0-based).  Clamp before subtracting.
-        let effective_x = self.cursor.pos.x.min(self.width.saturating_sub(1));
-        let mut new_x = effective_x - 1;
-
-        // Skip left over continuation cells of a wide glyph
-        while new_x > 0 {
-            if let Some(cell) = row.get_char_at(new_x) {
-                if !cell.is_continuation() {
-                    break;
-                }
-            } else {
-                break;
-            }
-            new_x -= 1;
-        }
-
-        self.cursor.pos.x = new_x;
-
-        // This MUST be the only post-condition for backspace.
-        debug_assert!(self.cursor.pos.y < self.rows.len());
-        debug_assert!(self.cursor.pos.x < self.width);
-    }
-
-    /// Move the cursor up one row and to the last column (reverse wrap).
-    ///
-    /// If the cursor is already at the top of the visible screen and
-    /// `into_scrollback` is true, the cursor enters the scrollback region.
-    /// If at the top and `into_scrollback` is false, the cursor stays put.
-    fn reverse_wrap_up(&mut self, into_scrollback: XtRevWrap2) {
-        let visible_start = self.visible_window_start(0);
-
-        if self.cursor.pos.y > visible_start {
-            // Not at the top of the visible screen — wrap to previous row.
-            self.cursor.pos.y -= 1;
-            self.cursor.pos.x = self.width.saturating_sub(1);
-        } else if into_scrollback == XtRevWrap2::Enabled && self.cursor.pos.y > 0 {
-            // At top of visible screen, but scrollback exists and
-            // extended reverse-wrap is enabled.
-            self.cursor.pos.y -= 1;
-            self.cursor.pos.x = self.width.saturating_sub(1);
-        }
-        // Otherwise: at absolute top or scrollback not allowed — no-op.
-    }
-
-    /// Fill the entire visible screen with 'E' characters (DECALN — ESC # 8).
-    ///
-    /// Also resets the scroll region to full screen and moves the cursor
-    /// to the home position (0, 0).
-    pub fn screen_alignment_test(&mut self) {
-        use freminal_common::buffer_states::format_tag::FormatTag;
-        use freminal_common::buffer_states::tchar::TChar;
-
-        let visible_start = self.visible_window_start(0);
-        let visible_end = visible_start + self.height;
-
-        // Ensure we have enough rows
-        while self.rows.len() < visible_end {
-            self.rows.push(crate::row::Row::new(self.width));
-            self.row_cache.push(None);
-        }
-
-        let default_tag = FormatTag::default();
-        let e_chars: Vec<TChar> = vec![TChar::Ascii(b'E'); self.width];
-        for i in visible_start..visible_end.min(self.rows.len()) {
-            self.image_cell_count -= self.rows[i].count_image_cells();
-            self.rows[i].clear();
-            self.rows[i].insert_text(0, &e_chars, &default_tag);
-            // Invalidate row cache
-            if i < self.row_cache.len() {
-                self.row_cache[i] = None;
-            }
-        }
-
-        // Reset scroll region to full screen
-        self.scroll_region_top = 0;
-        self.scroll_region_bottom = self.height.saturating_sub(1);
-
-        // Move cursor to home position
-        self.cursor.pos.x = 0;
-        self.cursor.pos.y = visible_start;
-    }
-
-    /// Handle ANSI LF (line feed), IND-style advance, and LNM mode.
-    ///
-    /// Moves the cursor down one row within the current `DECSTBM` scroll region.
-    /// When the cursor is at the bottom margin the region scrolls up by one line.
-    /// In `LNM` (new-line) mode an implicit CR is also applied.
-    pub fn handle_lf(&mut self) {
-        // Clear the implicit pending-wrap state.  When the cursor is at
-        // `x == width` (one past the last column) it means a character was
-        // just written at the rightmost column and the next printable
-        // character should wrap.  LF must cancel that deferred wrap —
-        // the cursor stays at the last column of the current/new row
-        // rather than wrapping on the next character write.
-        if self.cursor.pos.x >= self.width {
-            self.cursor.pos.x = self.width.saturating_sub(1);
-        }
-
-        match self.kind {
-            BufferType::Primary => {
-                // LNM: CR implied
-                if self.lnm_enabled == Lnm::NewLine {
-                    self.cursor.pos.x = 0;
-                }
-
-                let sy = self.cursor_screen_y();
-
-                // --- DECSTBM applies ONLY when:
-                //     - region is not full screen
-                //     - not scrolled back (already ensured above)
-                if self.scroll_region_top == 0 && self.scroll_region_bottom == self.height - 1 {
-                    //
-                    // FAST PATH: FULL-SCREEN REGION
-                    //
-                    // When the cursor is above the last visible row we simply move
-                    // it down.  We must NOT clear the destination row if it already
-                    // holds real content (e.g. logo lines written in a previous pass
-                    // before a CUU brought the cursor back up).  We only clear when
-                    // the destination is a pristine ScrollFill placeholder, and we
-                    // only scroll-in a brand-new blank row when the cursor is at the
-                    // very bottom of the visible window.
-                    self.cursor.pos.y += 1;
-
-                    if self.cursor.pos.y >= self.rows.len() {
-                        // Row doesn't exist yet — always create it fresh.
-                        // BCE: fill with current SGR background.
-                        let mut new_row = Row::new_with_origin(
-                            self.width,
-                            RowOrigin::HardBreak,
-                            RowJoin::NewLogicalLine,
-                        );
-                        new_row.fill_with_tag(&self.current_tag);
-                        self.rows.push(new_row);
-                        self.row_cache.push(None);
-                    } else {
-                        let row = &mut self.rows[self.cursor.pos.y];
-                        if row.origin == RowOrigin::ScrollFill {
-                            // Pristine placeholder: stamp it as a real hard-break
-                            // line but leave its (empty) cell content alone.
-                            row.origin = RowOrigin::HardBreak;
-                            row.join = RowJoin::NewLogicalLine;
-                        } else if sy == self.height.saturating_sub(1) {
-                            // Cursor was at the bottom of the visible window and the
-                            // next slot already has content from old scrollback — this
-                            // is the newly-scrolled-in line, so wipe it.  Use default
-                            // background (no BCE) — same rationale as `push_row`.
-                            self.image_cell_count -= row.count_image_cells();
-                            row.origin = RowOrigin::HardBreak;
-                            row.join = RowJoin::NewLogicalLine;
-                            row.clear();
-                        }
-                        // Otherwise (cursor was above the bottom, row has real
-                        // content): leave the row completely untouched.
-                    }
-
-                    // PTY always at scroll_offset=0; return value is always 0 here.
-                    let _ = self.enforce_scrollback_limit(0);
-                    self.debug_assert_invariants();
-                    return;
-                }
-
-                //
-                // PARTIAL REGION (true DECSTBM behavior)
-                //
-                if sy >= self.scroll_region_top && sy <= self.scroll_region_bottom {
-                    if sy < self.scroll_region_bottom {
-                        // Move cursor down inside region
-                        self.cursor.pos.y += 1;
-                        // If the row doesn't exist yet (buffer filling up), create it.
-                        while self.cursor.pos.y >= self.rows.len() {
-                            self.push_row(RowOrigin::HardBreak, RowJoin::NewLogicalLine);
-                        }
-                    } else {
-                        // At bottom margin → scroll region UP
-                        self.scroll_region_up_primary();
-                        // cursor stays in place
-                    }
-                } else {
-                    // Outside DECSTBM → just move down with no scrolling
-                    if self.cursor.pos.y + 1 < self.rows.len() {
-                        self.cursor.pos.y += 1;
-                    } else {
-                        let mut new_row = Row::new_with_origin(
-                            self.width,
-                            RowOrigin::HardBreak,
-                            RowJoin::NewLogicalLine,
-                        );
-                        new_row.fill_with_tag(&self.current_tag);
-                        self.rows.push(new_row);
-                        self.row_cache.push(None);
-                        self.cursor.pos.y = self.rows.len() - 1;
-                    }
-                }
-
-                // PTY always at scroll_offset=0; return value is always 0 here.
-                let _ = self.enforce_scrollback_limit(0);
-                self.debug_assert_invariants();
-            }
-
-            BufferType::Alternate => {
-                if self.lnm_enabled == Lnm::NewLine {
-                    self.cursor.pos.x = 0;
-                }
-
-                let y = self.cursor.pos.y;
-
-                if y >= self.scroll_region_top && y <= self.scroll_region_bottom {
-                    if y < self.scroll_region_bottom {
-                        if self.cursor.pos.y + 1 < self.height {
-                            self.cursor.pos.y += 1;
-                        }
-                    } else {
-                        self.scroll_slice_up(self.scroll_region_top, self.scroll_region_bottom);
-                    }
-                } else if self.cursor.pos.y + 1 < self.height {
-                    self.cursor.pos.y += 1;
-                }
-
-                self.debug_assert_invariants();
-            }
-        }
-    }
-
-    /// Handle ANSI CR (carriage return) — move cursor to column 0.
-    pub const fn handle_cr(&mut self) {
-        self.cursor.pos.x = 0;
-    }
-
-    /// IND – Index (move down, scroll within DECSTBM region).
-    /// Same as LF, but *does not* honor LNM (no implicit CR).
-    pub fn handle_ind(&mut self) {
-        // Temporarily disable LNM so `handle_lf` won't do CR.
-        let old_lnm = self.lnm_enabled;
-        self.lnm_enabled = Lnm::LineFeed;
-        self.handle_lf();
-        self.lnm_enabled = old_lnm;
-    }
-
-    /// NEL – Next Line (CR + LF with scrolling in DECSTBM region).
-    pub fn handle_nel(&mut self) {
-        // Explicit CR then LF – this is allowed to honor LNM.
-        self.handle_cr();
-        self.handle_lf();
-    }
-
-    /// RI – Reverse Index.
-    /// Move the cursor up; at the top margin of DECSTBM region,
-    /// scroll the region down by one line (blank line at top).
-    pub fn handle_ri(&mut self) {
-        // Clear implicit pending-wrap state (same as handle_lf).
-        if self.cursor.pos.x >= self.width {
-            self.cursor.pos.x = self.width.saturating_sub(1);
-        }
-
-        match self.kind {
-            BufferType::Primary => {
-                let sy = self.cursor_screen_y();
-
-                if sy >= self.scroll_region_top && sy <= self.scroll_region_bottom {
-                    if sy > self.scroll_region_top {
-                        // move up inside region
-                        self.cursor.pos.y -= 1;
-                    } else {
-                        // at top margin → scroll region DOWN
-                        self.scroll_region_down_primary();
-                    }
-                } else {
-                    // outside region → never scroll
-                    if self.cursor.pos.y > 0 {
-                        self.cursor.pos.y -= 1;
-                    }
-                }
-
-                self.debug_assert_invariants();
-            }
-
-            BufferType::Alternate => {
-                let y = self.cursor.pos.y;
-
-                if y >= self.scroll_region_top && y <= self.scroll_region_bottom {
-                    if y > self.scroll_region_top {
-                        self.cursor.pos.y -= 1;
-                    } else {
-                        self.scroll_slice_down(self.scroll_region_top, self.scroll_region_bottom);
-                    }
-                } else if self.cursor.pos.y > 0 {
-                    self.cursor.pos.y -= 1;
-                }
-                self.debug_assert_invariants();
-            }
-        }
-    }
-
-    /// IL – Insert Lines within DECSTBM region.
-    /// Insert `n` blank lines at the cursor row, shifting lines down and
-    /// discarding at the bottom of the region.
-    pub fn insert_lines(&mut self, n: usize) {
-        if n == 0 {
-            return;
-        }
-
-        match self.kind {
-            BufferType::Alternate => {
-                // leave as-is
-                let y = self.cursor.pos.y;
-                if y < self.scroll_region_top || y > self.scroll_region_bottom {
-                    return;
-                }
-
-                let max_lines = self.scroll_region_bottom.saturating_sub(y) + 1;
-                let count = n.min(max_lines);
-
-                if self.declrmm_enabled == Declrmm::Enabled {
-                    let (left, right) = (self.scroll_region_left, self.scroll_region_right);
-                    for _ in 0..count {
-                        self.scroll_slice_down_columns(y, self.scroll_region_bottom, left, right);
-                    }
-                } else {
-                    for _ in 0..count {
-                        self.scroll_slice_down(y, self.scroll_region_bottom);
-                    }
-                }
-
-                self.debug_assert_invariants();
-            }
-
-            BufferType::Primary => {
-                let sy = self.cursor_screen_y();
-                if sy < self.scroll_region_top || sy > self.scroll_region_bottom {
-                    return;
-                }
-
-                let (t, b) = self.scroll_region_rows();
-                let offset = sy - self.scroll_region_top;
-                let row = t + offset;
-
-                let count = n.min(b - row + 1);
-                if self.declrmm_enabled == Declrmm::Enabled {
-                    let (left, right) = (self.scroll_region_left, self.scroll_region_right);
-                    for _ in 0..count {
-                        self.scroll_slice_down_columns(row, b, left, right);
-                    }
-                } else {
-                    for _ in 0..count {
-                        self.scroll_slice_down(row, b);
-                    }
-                }
-
-                self.debug_assert_invariants();
-            }
-        }
-    }
-
-    /// DL – Delete Lines within DECSTBM region.
-    /// Delete `n` lines at the cursor row, shifting lines up and
-    /// inserting blanks at the bottom of the region.
-    pub fn delete_lines(&mut self, n: usize) {
-        if n == 0 {
-            return;
-        }
-
-        match self.kind {
-            BufferType::Alternate => {
-                let y = self.cursor.pos.y;
-                if y < self.scroll_region_top || y > self.scroll_region_bottom {
-                    return;
-                }
-
-                let max_lines = self.scroll_region_bottom.saturating_sub(y) + 1;
-                let count = n.min(max_lines);
-
-                if self.declrmm_enabled == Declrmm::Enabled {
-                    let (left, right) = (self.scroll_region_left, self.scroll_region_right);
-                    for _ in 0..count {
-                        self.scroll_slice_up_columns(y, self.scroll_region_bottom, left, right);
-                    }
-                } else {
-                    for _ in 0..count {
-                        self.scroll_slice_up(y, self.scroll_region_bottom);
-                    }
-                }
-
-                self.debug_assert_invariants();
-            }
-
-            BufferType::Primary => {
-                let sy = self.cursor_screen_y();
-                if sy < self.scroll_region_top || sy > self.scroll_region_bottom {
-                    return;
-                }
-
-                let (t, b) = self.scroll_region_rows();
-                let offset = sy - self.scroll_region_top;
-                let row = t + offset;
-
-                let count = n.min(b - row + 1);
-                if self.declrmm_enabled == Declrmm::Enabled {
-                    let (left, right) = (self.scroll_region_left, self.scroll_region_right);
-                    for _ in 0..count {
-                        self.scroll_slice_up_columns(row, b, left, right);
-                    }
-                } else {
-                    for _ in 0..count {
-                        self.scroll_slice_up(row, b);
-                    }
-                }
-
-                self.debug_assert_invariants();
-            }
-        }
-    }
-
-    /// Implements DECSC – Save Cursor.
-    ///
-    /// Saves the current cursor position (and associated `CursorState`).
-    pub fn save_cursor(&mut self) {
-        self.saved_cursor = Some(self.cursor.clone());
-    }
-
-    /// Implements DECRC – Restore Cursor.
-    ///
-    /// Restores the previously saved cursor position.  If no cursor has been
-    /// saved, this is a no-op.  The restored position is clamped to the
-    /// current buffer dimensions so a resize between save and restore never
-    /// produces an out-of-bounds cursor.
-    pub fn restore_cursor(&mut self) {
-        if let Some(saved) = self.saved_cursor.clone() {
-            self.cursor = saved;
-            // Clamp to current dimensions after restore.
-            if self.width > 0 {
-                self.cursor.pos.x = self.cursor.pos.x.min(self.width - 1);
-            }
-            let max_row = self.rows.len().saturating_sub(1);
-            self.cursor.pos.y = self.cursor.pos.y.min(max_row);
-            self.debug_assert_invariants();
-        }
-        // No saved cursor → silent no-op.
-    }
-
-    /// Implements ICH – Insert Characters (spaces).
-    pub fn insert_spaces(&mut self, n: usize) {
-        let row = self.cursor.pos.y;
-        let col = self.cursor.pos.x;
-
-        if row >= self.rows.len() {
-            return;
-        }
-
-        // Sweep any image cells in the range that will be shifted off the right
-        // edge before the insert happens.  Cells shifted past the right margin
-        // are discarded, so we must update image_cell_count now.
-        let remaining_images = if self.image_cell_count > 0 {
-            let right = if self.declrmm_enabled == Declrmm::Enabled {
-                self.scroll_region_right + 1
-            } else {
-                self.width
-            };
-            self.collect_and_clear_image_ids_in_rows(row, row + 1, Some(col), Some(right));
-            // Count Kitty images that survived the sweep in the cells that
-            // will overflow past the right edge when we insert `n` blanks.
-            let overflow_start = right.saturating_sub(n).max(col);
-            self.rows[row].count_image_cells_in_range(overflow_start, right)
-        } else {
-            0
-        };
-
-        let tag = self.current_tag.clone();
-        if self.declrmm_enabled == Declrmm::Enabled {
-            // ICH: shift only within [col, scroll_region_right]; cells that
-            // fall off the right margin are discarded.
-            self.rows[row].insert_spaces_at_with_right_limit(
-                col,
-                n,
-                &tag,
-                self.scroll_region_right + 1,
-            );
-        } else {
-            self.rows[row].insert_spaces_at(col, n, &tag);
-        }
-
-        self.image_cell_count -= remaining_images;
-        // Cursor does NOT move for ICH
-    }
-
-    /// Implements DCH – Delete Characters.
-    ///
-    /// Removes `n` cells starting at the cursor column on the cursor row, shifting
-    /// the cells to the right of the deleted range left to fill the gap. Cells
-    /// shifted off the right edge are discarded.  The cursor does not move.
-    ///
-    /// When DECLRMM is active, the operation is confined to
-    /// `[col, scroll_region_right]`; cells outside the right margin are
-    /// unaffected.
-    ///
-    /// Wide-glyph cleanup is delegated to [`Row::delete_cells_at`] /
-    /// [`Row::delete_cells_at_with_right_limit`].
-    pub fn delete_chars(&mut self, n: usize) {
-        let row = self.cursor.pos.y;
-        let col = self.cursor.pos.x;
-
-        if row >= self.rows.len() {
-            // Row doesn't even exist — nothing to delete.
-            return;
-        }
-
-        // Sweep image cells in the range being deleted before the row-level
-        // operation removes them without updating image_cell_count.
-        let remaining_images = if self.image_cell_count > 0 {
-            let end_col = if self.declrmm_enabled == Declrmm::Enabled {
-                (col + n).min(self.scroll_region_right + 1)
-            } else {
-                col + n
-            };
-            self.collect_and_clear_image_ids_in_rows(row, row + 1, Some(col), Some(end_col));
-            // Count Kitty images that survived the sweep in the deleted range.
-            self.rows[row].count_image_cells_in_range(col, end_col)
-        } else {
-            0
-        };
-
-        if self.declrmm_enabled == Declrmm::Enabled {
-            self.rows[row].delete_cells_at_with_right_limit(
-                col,
-                n,
-                self.scroll_region_right + 1,
-                &self.current_tag.clone(),
-            );
-        } else {
-            let tag = self.current_tag.clone();
-            self.rows[row].delete_cells_at(col, n, &tag);
-        }
-
-        self.image_cell_count -= remaining_images;
-        self.debug_assert_invariants();
-    }
-
-    /// Implements ECH – Erase Characters.
-    ///
-    /// Replaces `n` cells starting at the cursor column with blanks using the current
-    /// format tag.  Cells to the right of the erased range are **not** shifted.
-    /// The cursor does not move.
-    ///
-    /// When DECLRMM is active, `n` is clamped so the erase never crosses the
-    /// right margin.
-    ///
-    /// Wide-glyph cleanup is delegated to [`Row::erase_cells_at`].
-    pub fn erase_chars(&mut self, n: usize) {
-        let row = self.cursor.pos.y;
-        let col = self.cursor.pos.x;
-
-        if row >= self.rows.len() {
-            // Row doesn't exist yet — nothing to erase.
-            return;
-        }
-
-        // Clamp erase count to right margin when DECLRMM is active.
-        let effective_n = if self.declrmm_enabled == Declrmm::Enabled {
-            let max_erase = self.scroll_region_right.saturating_sub(col) + 1;
-            n.min(max_erase)
-        } else {
-            n
-        };
-
-        // Sweep image cells that are about to be erased (non-Kitty images).
-        self.collect_and_clear_image_ids_in_rows(row, row + 1, Some(col), Some(col + effective_n));
-
-        // Count remaining image cells (e.g. Kitty) in the range being erased.
-        let remaining = if self.image_cell_count > 0 {
-            self.rows[row].count_image_cells_in_range(col, col + effective_n)
-        } else {
-            0
-        };
-
-        let tag = self.current_tag.clone();
-        self.rows[row].erase_cells_at(col, effective_n, &tag);
-        self.image_cell_count -= remaining;
-
-        self.debug_assert_invariants();
-    }
-
-    fn reset_scroll_region_to_full(&mut self) {
-        self.scroll_region_top = 0;
-        self.scroll_region_bottom = self.height.saturating_sub(1);
-        // Reset cursor to home position (screen row 0, col 0).
-        // Use set_cursor_pos so rows are created if they don't exist yet.
-        self.set_cursor_pos(Some(0), Some(0));
-    }
-
-    /// Set DECSTBM scroll region (1-based inclusive).
-    /// If invalid, resets to full screen.
-    pub fn set_scroll_region(&mut self, top1: usize, bottom1: usize) {
-        // 0 or missing → ignore and reset
-        if top1 == 0 || bottom1 == 0 {
-            self.reset_scroll_region_to_full();
-            return;
-        }
-
-        // Convert to 0-based
-        let top = top1.saturating_sub(1);
-        let bottom = bottom1.saturating_sub(1);
-
-        // Validate
-        if top >= bottom || bottom >= self.height {
-            self.reset_scroll_region_to_full();
-            return;
-        }
-
-        self.scroll_region_top = top;
-        self.scroll_region_bottom = bottom;
-
-        // DECSTBM always homes the cursor to (0, 0).
-        // When DECOM is enabled, set_cursor_pos interprets y=0 as
-        // scroll_region_top.  When DECOM is disabled, y=0 is screen top.
-        // Both cases are handled correctly by a single call.
-        self.set_cursor_pos(Some(0), Some(0));
-    }
-
-    /// Return the current scroll region as 0-based inclusive `(top, bottom)`.
-    #[must_use]
-    pub const fn scroll_region(&self) -> (usize, usize) {
-        (self.scroll_region_top, self.scroll_region_bottom)
-    }
-
-    /// Reset left/right margins to the full terminal width.
-    /// Public so `TerminalHandler` can call it when DECLRMM is reset.
-    pub const fn reset_scroll_region_left_right(&mut self) {
-        self.scroll_region_left = 0;
-        self.scroll_region_right = self.width.saturating_sub(1);
-    }
-
-    /// Enable or disable DECLRMM (`?69`).
-    ///
-    /// When disabled, left/right margins are also reset to full-screen.
-    pub fn set_declrmm(&mut self, mode: Declrmm) {
-        self.declrmm_enabled = mode;
-        if self.declrmm_enabled == Declrmm::Disabled {
-            self.reset_scroll_region_left_right();
-        }
-    }
-
-    /// Returns the current DECLRMM (`?69`) state.
-    #[must_use]
-    pub const fn is_declrmm_enabled(&self) -> Declrmm {
-        self.declrmm_enabled
-    }
-
-    /// Set DECSLRM left/right margins (1-based inclusive).
-    ///
-    /// Only effective when DECLRMM (`?69`) is active; the caller
-    /// (`TerminalHandler::handle_set_left_right_margins`) must gate on that flag.
-    /// If parameters are invalid, resets to full-width.
-    pub fn set_left_right_margins(&mut self, left1: usize, right1: usize) {
-        // 0 → treat as 1 (reset)
-        if left1 == 0 || right1 == 0 {
-            self.reset_scroll_region_left_right();
-            self.set_cursor_pos(Some(0), Some(0));
-            return;
-        }
-
-        let left = left1.saturating_sub(1);
-        let right = right1.saturating_sub(1);
-
-        if left >= right || right >= self.width {
-            self.reset_scroll_region_left_right();
-            self.set_cursor_pos(Some(0), Some(0));
-            return;
-        }
-
-        self.scroll_region_left = left;
-        self.scroll_region_right = right;
-
-        // DECSLRM homes the cursor to (col 0, row 0) — just like DECSTBM.
-        self.set_cursor_pos(Some(0), Some(0));
-    }
-
-    /// Return the current left/right margins as 0-based inclusive `(left, right)`.
-    #[must_use]
-    pub const fn left_right_margins(&self) -> (usize, usize) {
-        (self.scroll_region_left, self.scroll_region_right)
-    }
-
-    /// Index in `rows` of the first visible line for the given `scroll_offset`.
-    /// The PTY thread always calls this with `scroll_offset = 0`; the GUI thread
-    /// passes the value from `ViewState`.
-    #[must_use]
-    fn visible_window_start(&self, scroll_offset: usize) -> usize {
-        if self.rows.is_empty() || self.height == 0 {
-            return 0;
-        }
-
-        let total = self.rows.len();
-        let h = self.height.min(total);
-        let offset = scroll_offset.min(self.max_scroll_offset());
-
-        total.saturating_sub(h + offset)
-    }
-
-    /// Return `true` if any row in the visible window is dirty (needs re-flattening).
-    ///
-    /// The PTY thread calls this with `scroll_offset = 0`.  A `false` result means
-    /// the cached flat representation for every visible row is still valid, so
-    /// `build_snapshot` can skip the flatten step entirely and reuse the previous
-    /// `visible_chars` / `visible_tags` vectors.
-    #[must_use]
-    pub fn any_visible_dirty(&self, scroll_offset: usize) -> bool {
-        if self.rows.is_empty() || self.height == 0 {
-            return false;
-        }
-        let vis_start = self.visible_window_start(scroll_offset);
-        let vis_end = (vis_start + self.height).min(self.rows.len());
-        self.rows[vis_start..vis_end].iter().any(|r| r.dirty)
-    }
-
-    /// Extract image placements for all cells in the visible window.
-    ///
-    /// Returns a flat `Vec` of `Option<ImagePlacement>`, one entry per cell,
-    /// in row-major order (row 0, col 0..width; row 1, col 0..width; …).
-    ///
-    /// `None` means the cell carries no image data.
-    /// `Some(placement)` means the cell is part of an inline image.
-    ///
-    /// The length of the returned `Vec` is `height * width` (clamped to the
-    /// actual number of visible rows × terminal width), matching the layout
-    /// of `visible_chars` so the caller can index them in parallel.
-    #[must_use]
-    pub fn visible_image_placements(&self, scroll_offset: usize) -> Vec<Option<ImagePlacement>> {
-        if self.rows.is_empty() || self.height == 0 || self.width == 0 {
-            return Vec::new();
-        }
-        let vis_start = self.visible_window_start(scroll_offset);
-        let vis_end = (vis_start + self.height).min(self.rows.len());
-        let mut out = Vec::with_capacity((vis_end - vis_start) * self.width);
-        for row in &self.rows[vis_start..vis_end] {
-            let cells = row.cells();
-            for col in 0..self.width {
-                let placement = cells.get(col).and_then(|c| c.image_placement()).cloned();
-                out.push(placement);
-            }
-        }
-        out
-    }
-
-    /// Returns `true` if any cell in the visible window carries an image
-    /// placement.  Used by `build_snapshot` to cheaply decide whether to
-    /// include image data in the snapshot.
-    ///
-    /// Short-circuits in O(1) when the buffer has no image cells at all
-    /// (the overwhelmingly common case).
-    #[must_use]
-    pub fn has_visible_images(&self, scroll_offset: usize) -> bool {
-        // Fast path: no image cells anywhere in the buffer.
-        if self.image_cell_count == 0 {
-            return false;
-        }
-
-        if self.rows.is_empty() || self.height == 0 {
-            return false;
-        }
-        let vis_start = self.visible_window_start(scroll_offset);
-        let vis_end = (vis_start + self.height).min(self.rows.len());
-        self.rows[vis_start..vis_end]
-            .iter()
-            .any(|r| r.cells().iter().any(|c| c.image_placement().is_some()))
-    }
-
-    /// Cursor Y expressed in "screen coordinates" (0..height-1).
-    /// If the buffer is shorter than the height, we just return the raw Y.
-    /// Always computed relative to the live bottom (`scroll_offset` = 0), because the
-    /// PTY thread only ever mutates the buffer at the live bottom.
-    fn cursor_screen_y(&self) -> usize {
-        if self.rows.is_empty() || self.height == 0 {
-            return 0;
-        }
-
-        let start = self.visible_window_start(0);
-        self.cursor.pos.y.saturating_sub(start)
-    }
-
-    /// Convert DECSTBM region (screen coords) into buffer row indices (rows[])
-    ///
-    /// Ensures `self.rows` is extended to at least `height` entries so that
-    /// the returned indices always point to real rows.  Without this, an early
-    /// buffer (`rows.len()` < height) would clamp both top and bottom to the
-    /// same index, causing every scroll operation to silently no-op.
-    fn scroll_region_rows(&mut self) -> (usize, usize) {
-        let start = self.visible_window_start(0);
-        let required = start + self.scroll_region_bottom + 1;
-        while self.rows.len() < required {
-            self.push_row(RowOrigin::ScrollFill, RowJoin::NewLogicalLine);
-        }
-        let top = start + self.scroll_region_top;
-        let bottom = start + self.scroll_region_bottom;
-        (top, bottom)
-    }
-
-    /// Scroll DECSTBM region UP by 1 (primary buffer)
-    fn scroll_region_up_primary(&mut self) {
-        let (t, b) = self.scroll_region_rows();
-        if t < b {
-            self.scroll_slice_up(t, b);
-        }
-    }
-
-    /// Scroll DECSTBM region DOWN by 1 (primary buffer)
-    fn scroll_region_down_primary(&mut self) {
-        let (t, b) = self.scroll_region_rows();
-        if t < b {
-            self.scroll_slice_down(t, b);
-        }
-    }
-
-    /// Check whether the cursor is at the bottom margin of the DECSTBM
-    /// scroll region.  Used by `insert_text` to decide whether a right-margin
-    /// wrap should scroll the region instead of advancing past it.
-    fn is_cursor_at_scroll_region_bottom(&self) -> bool {
-        match self.kind {
-            BufferType::Primary => {
-                let sy = self.cursor_screen_y();
-                sy == self.scroll_region_bottom
-            }
-            BufferType::Alternate => self.cursor.pos.y == self.scroll_region_bottom,
-        }
-    }
-
-    /// Scroll the DECSTBM region up by one line during an autowrap at the
-    /// bottom margin.  Handles both Primary (with scrollback) and Alternate
-    /// (fixed grid) buffer types.
-    fn scroll_region_up_for_wrap(&mut self) {
-        match self.kind {
-            BufferType::Primary => {
-                self.scroll_region_up_primary();
-            }
-            BufferType::Alternate => {
-                if self.scroll_region_top < self.scroll_region_bottom {
-                    self.scroll_slice_up(self.scroll_region_top, self.scroll_region_bottom);
-                }
-            }
-        }
-    }
-
-    /// SU — Scroll the scroll region UP by `n` lines.
-    /// Content moves up; blank lines appear at the bottom.
-    /// If no scroll region is set, operates on the whole screen.
-    pub fn scroll_region_up_n(&mut self, n: usize) {
-        let (t, b) = self.scroll_region_rows();
-        if t >= b {
-            return;
-        }
-        // Region indices are inclusive, so region has (b - t + 1) rows.
-        let region_size = b - t + 1;
-        let clamped = n.min(region_size);
-        for _ in 0..clamped {
-            self.scroll_slice_up(t, b);
-        }
-    }
-
-    /// SD — Scroll the scroll region DOWN by `n` lines.
-    /// Content moves down; blank lines appear at the top.
-    /// If no scroll region is set, operates on the whole screen.
-    pub fn scroll_region_down_n(&mut self, n: usize) {
-        let (t, b) = self.scroll_region_rows();
-        if t >= b {
-            return;
-        }
-        // Region indices are inclusive, so region has (b - t + 1) rows.
-        let region_size = b - t + 1;
-        let clamped = n.min(region_size);
-        for _ in 0..clamped {
-            self.scroll_slice_down(t, b);
-        }
-    }
-
-    /// Scroll a contiguous vertical slice [first, last] UP by one line.
-    /// Rows outside that range are untouched. New bottom line is blank.
-    fn scroll_slice_up(&mut self, first: usize, last: usize) {
-        if first >= last {
-            return;
-        }
-        if last >= self.rows.len() {
-            return;
-        }
-
-        for row_idx in first..last {
-            let next = self.rows[row_idx + 1].clone();
-            self.rows[row_idx] = next;
-            // Rotate the cache entry in lockstep: a moved row keeps its cached
-            // flat representation (it hasn't changed content, only position).
-            self.row_cache[row_idx] = self.row_cache[row_idx + 1].take();
-        }
-
-        // The original rows[last] was not shifted (the loop only copies
-        // rows[row_idx+1] into rows[row_idx] for row_idx in first..last).
-        // It is now replaced with a blank row; deduct any image cells it held.
-        self.image_cell_count -= self.rows[last].count_image_cells();
-        let new_row = Row::new(self.width);
-        // Scroll-created blank rows use default background (no BCE).
-        // See `push_row` comment for rationale.
-        self.rows[last] = new_row;
-        // New blank row at `last` — no cached representation yet.
-        self.row_cache[last] = None;
-    }
-
-    /// Scroll a contiguous vertical slice [first, last] DOWN by one line.
-    /// Rows outside that range are untouched. New top line is blank.
-    fn scroll_slice_down(&mut self, first: usize, last: usize) {
-        if first >= last {
-            return;
-        }
-        if last >= self.rows.len() {
-            return;
-        }
-
-        for row_idx in (first + 1..=last).rev() {
-            let prev = self.rows[row_idx - 1].clone();
-            self.rows[row_idx] = prev;
-            // Rotate the cache entry in lockstep.
-            self.row_cache[row_idx] = self.row_cache[row_idx - 1].take();
-        }
-
-        // The original rows[first] was not shifted (the loop only copies
-        // rows[row_idx-1] into rows[row_idx] for row_idx in first+1..=last).
-        // It is now replaced with a blank row; deduct any image cells it held.
-        self.image_cell_count -= self.rows[first].count_image_cells();
-        let new_row = Row::new(self.width);
-        // Scroll-created blank rows use default background (no BCE).
-        // See `push_row` comment for rationale.
-        self.rows[first] = new_row;
-        // New blank row at `first` — no cached representation yet.
-        self.row_cache[first] = None;
-    }
-
-    /// Column-selective scroll-up: shifts cells within `[left_col, right_col]`
-    /// on rows `[first, last]` up by one, without touching cells outside that
-    /// horizontal range.  Used by `insert_lines` / `delete_lines` when DECLRMM
-    /// is active.
-    fn scroll_slice_up_columns(
-        &mut self,
-        first: usize,
-        last: usize,
-        left_col: usize,
-        right_col: usize,
-    ) {
-        if first >= last || last >= self.rows.len() || left_col > right_col {
-            return;
-        }
-        // Sweep image cells in the destination columns [left_col, right_col] across
-        // all rows [first, last] before any cells are overwritten or erased.
-        // This covers both the copy-overwrite path (rows first..last) and the
-        // final erase on rows[last].
-        //
-        // After the sweep, count remaining image cells (e.g. Kitty) in the
-        // affected range before the copy/erase so we can compute the net
-        // change after the operation.
-        let images_before = if self.image_cell_count > 0 {
-            self.collect_and_clear_image_ids_in_rows(
-                first,
-                last + 1,
-                Some(left_col),
-                Some(right_col + 1),
-            );
-            (first..=last)
-                .map(|i| self.rows[i].count_image_cells_in_range(left_col, right_col + 1))
-                .sum::<usize>()
-        } else {
-            0
-        };
-        let tag = self.current_tag.clone();
-        for row_idx in first..last {
-            // Copy cells [left_col, right_col] from row_idx+1 into row_idx.
-            // Use resolve_cell to handle sparse rows correctly — columns
-            // beyond the stored cell count are treated as implicit blanks.
-            let src_cells: Vec<_> = (left_col..=right_col)
-                .map(|col| self.rows[row_idx + 1].resolve_cell(col))
-                .collect();
-            let row = &mut self.rows[row_idx];
-            // Ensure storage.
-            if row.cells_mut().len() < right_col + 1 {
-                let width = row.width();
-                while row.cells_mut().len() < (right_col + 1).min(width) {
-                    row.cells_mut_push(Cell::blank_with_tag(FormatTag::default()));
-                }
-            }
-            for (offset, cell) in src_cells.into_iter().enumerate() {
-                let dst = left_col + offset;
-                if dst <= right_col && dst < row.cells().len() {
-                    row.cells_mut()[dst] = cell;
-                }
-            }
-            row.mark_dirty();
-            self.row_cache[row_idx] = None;
-        }
-        // Blank [left_col, right_col] on the last row.
-        let row = &mut self.rows[last];
-        row.erase_cells_at(left_col, right_col - left_col + 1, &tag);
-        self.row_cache[last] = None;
-
-        // Adjust image_cell_count for any images lost during the shift/erase.
-        if images_before > 0 {
-            let images_after: usize = (first..=last)
-                .map(|i| self.rows[i].count_image_cells_in_range(left_col, right_col + 1))
-                .sum();
-            self.image_cell_count -= images_before.saturating_sub(images_after);
-        }
-    }
-
-    /// Column-selective scroll-down: shifts cells within `[left_col, right_col]`
-    /// on rows `[first, last]` down by one, without touching cells outside that
-    /// horizontal range.  Used by `insert_lines` / `delete_lines` when DECLRMM
-    /// is active.
-    fn scroll_slice_down_columns(
-        &mut self,
-        first: usize,
-        last: usize,
-        left_col: usize,
-        right_col: usize,
-    ) {
-        if first >= last || last >= self.rows.len() || left_col > right_col {
-            return;
-        }
-        // Sweep image cells in the destination columns [left_col, right_col] across
-        // all rows [first, last] before any cells are overwritten or erased.
-        // This covers both the copy-overwrite path (rows first+1..=last) and the
-        // final erase on rows[first].
-        let images_before = if self.image_cell_count > 0 {
-            self.collect_and_clear_image_ids_in_rows(
-                first,
-                last + 1,
-                Some(left_col),
-                Some(right_col + 1),
-            );
-            (first..=last)
-                .map(|i| self.rows[i].count_image_cells_in_range(left_col, right_col + 1))
-                .sum::<usize>()
-        } else {
-            0
-        };
-        let tag = self.current_tag.clone();
-        for row_idx in (first + 1..=last).rev() {
-            // Copy cells [left_col, right_col] from row_idx-1 into row_idx.
-            // Use resolve_cell to handle sparse rows correctly — columns
-            // beyond the stored cell count are treated as implicit blanks.
-            let src_cells: Vec<_> = (left_col..=right_col)
-                .map(|col| self.rows[row_idx - 1].resolve_cell(col))
-                .collect();
-            let row = &mut self.rows[row_idx];
-            if row.cells_mut().len() < right_col + 1 {
-                let width = row.width();
-                while row.cells_mut().len() < (right_col + 1).min(width) {
-                    row.cells_mut_push(Cell::blank_with_tag(FormatTag::default()));
-                }
-            }
-            for (offset, cell) in src_cells.into_iter().enumerate() {
-                let dst = left_col + offset;
-                if dst <= right_col && dst < row.cells().len() {
-                    row.cells_mut()[dst] = cell;
-                }
-            }
-            row.mark_dirty();
-            self.row_cache[row_idx] = None;
-        }
-        // Blank [left_col, right_col] on the first row.
-        let row = &mut self.rows[first];
-        row.erase_cells_at(left_col, right_col - left_col + 1, &tag);
-        self.row_cache[first] = None;
-
-        // Adjust image_cell_count for any images lost during the shift/erase.
-        if images_before > 0 {
-            let images_after: usize = (first..=last)
-                .map(|i| self.rows[i].count_image_cells_in_range(left_col, right_col + 1))
-                .sum();
-            self.image_cell_count -= images_before.saturating_sub(images_after);
-        }
-    }
-
-    // ----------------------------------------------------------
-    // Scrollback: only valid in the PRIMARY buffer
-    // ----------------------------------------------------------
-
-    /// How many lines above the live bottom the user can scroll.
-    #[must_use]
-    pub const fn max_scroll_offset(&self) -> usize {
-        if self.rows.len() <= self.height {
-            0
-        } else {
-            self.rows.len() - self.height
-        }
-    }
-
-    /// Compute a new scroll offset after scrolling upward by `lines`.
-    ///
-    /// Alternate buffer always returns 0 (no scrollback).
-    /// The caller is responsible for storing the returned value into `ViewState`.
-    #[must_use]
-    pub fn scroll_back(&self, scroll_offset: usize, lines: usize) -> usize {
-        if self.kind != BufferType::Primary {
-            return 0; // Alternate buffer: no scrollback
-        }
-
-        let max = self.max_scroll_offset();
-        if max == 0 {
-            return 0;
-        }
-
-        (scroll_offset + lines).min(max)
-    }
-
-    /// Compute a new scroll offset after scrolling downward by `lines`.
-    ///
-    /// The caller is responsible for storing the returned value into `ViewState`.
-    #[must_use]
-    pub fn scroll_forward(&self, scroll_offset: usize, lines: usize) -> usize {
-        if self.kind != BufferType::Primary {
-            return 0;
-        }
-
-        scroll_offset.saturating_sub(lines)
-    }
-
-    /// Returns `0` — the scroll offset for the live bottom view.
-    ///
-    /// Provided as a convenience so call sites read clearly.
-    #[must_use]
-    pub const fn scroll_to_bottom() -> usize {
-        0
-    }
-
-    /// Scroll the visible window up by one row, discarding the top row and appending a blank row at
-    /// the bottom.
-    ///
-    /// In the primary buffer the cursor row index is also decremented to follow the visible window.
-    pub fn scroll_up(&mut self) {
-        // Deduct any image cells in the row about to be removed.
-        self.image_cell_count -= self.rows[0].count_image_cells();
-        // remove topmost row (and its cache entry)
-        self.rows.remove(0);
-        self.row_cache.remove(0);
-
-        // add a new empty row at the bottom (BCE: fill with current SGR background)
-        let mut new_row = Row::new(self.width);
-        new_row.fill_with_tag(&self.current_tag);
-        self.rows.push(new_row);
-        self.row_cache.push(None);
-
-        // DO NOT move the cursor in alternate buffer
-        if self.kind == BufferType::Primary {
-            // primary buffer uses scrollback: move cursor with the visible window
-            if self.cursor.pos.y > 0 {
-                self.cursor.pos.y -= 1;
-            }
-        }
-    }
-
-    /// Move cursor to absolute position (CUP, HVP).
-    ///
-    /// `x` and `y` are 0-indexed screen coordinates.  `None` means "leave this
-    /// axis unchanged" (e.g. CHA only supplies x, VPA only supplies y, CUP
-    /// supplies both).
-    ///
-    /// When DECOM (origin mode) is enabled, `y` is relative to `scroll_region_top`
-    /// and is clamped to the scroll region height.  When DECOM is disabled, `y` is
-    /// relative to the top of the visible window and clamped to the screen height.
-    pub fn set_cursor_pos(&mut self, x: Option<usize>, y: Option<usize>) {
-        // `None` means "leave this axis unchanged" (e.g. CHA only supplies x,
-        // VPA only supplies y, CUP supplies both).
-        let new_x = match x {
-            Some(col) => col.min(self.width.saturating_sub(1)),
-            None => self.cursor.pos.x,
-        };
-
-        // y is a screen-relative coordinate (0 = top of visible window in normal
-        // mode, or 0 = top of scroll region in DECOM mode).
-        let new_buffer_y = match y {
-            Some(row) => {
-                if self.decom_enabled == Decom::OriginMode {
-                    // DECOM: row is relative to scroll_region_top, clamped to
-                    // the scroll region height.
-                    let region_height = self
-                        .scroll_region_bottom
-                        .saturating_sub(self.scroll_region_top);
-                    let clamped = row.min(region_height);
-                    let screen_row = self.scroll_region_top + clamped;
-                    self.visible_window_start(0) + screen_row
-                } else {
-                    let clamped = row.min(self.height.saturating_sub(1));
-                    // PTY always operates at live bottom (scroll_offset = 0)
-                    self.visible_window_start(0) + clamped
-                }
-            }
-            None => self.cursor.pos.y,
-        };
-
-        // Ensure rows exist up to the target position
-        while new_buffer_y >= self.rows.len() {
-            self.push_row(RowOrigin::ScrollFill, RowJoin::NewLogicalLine);
-        }
-
-        self.cursor.pos.x = new_x;
-        self.cursor.pos.y = new_buffer_y;
-        self.debug_assert_invariants();
-    }
-
-    /// Move cursor relatively (CUU, CUD, CUF, CUB)
-    /// Move the cursor by a relative offset `(dx, dy)` in screen coordinates.
-    ///
-    /// Positive `dx` moves right; negative moves left. Positive `dy` moves down; negative moves
-    /// up.  The cursor is clamped to the visible screen boundaries and never enters scrollback.
-    pub fn move_cursor_relative(&mut self, dx: i32, dy: i32) {
-        // When DECLRMM is active and moving horizontally, clamp to the
-        // left/right margins if the cursor is currently within the margin zone.
-        let new_x = if self.declrmm_enabled == Declrmm::Enabled && dx != 0 {
-            let cx = self.cursor.pos.x;
-            if cx >= self.scroll_region_left && cx <= self.scroll_region_right {
-                // Cursor is inside the margin zone: clamp to [left, right].
-                clamped_offset(cx, dx, self.scroll_region_left, self.scroll_region_right)
-            } else {
-                // Cursor is outside the margin zone: use normal full-width clamp.
-                clamped_offset(cx, dx, 0, self.width.saturating_sub(1))
-            }
-        } else {
-            clamped_offset(self.cursor.pos.x, dx, 0, self.width.saturating_sub(1))
-        };
-
-        let current_screen_y = self.cursor_screen_y();
-        let new_screen_y = clamped_offset(current_screen_y, dy, 0, self.height.saturating_sub(1));
-
-        // PTY always operates at live bottom (scroll_offset = 0)
-        let new_buffer_y = self.visible_window_start(0) + new_screen_y;
-
-        // Ensure rows exist
-        while new_buffer_y >= self.rows.len() {
-            self.push_row(RowOrigin::ScrollFill, RowJoin::NewLogicalLine);
-        }
-
-        self.cursor.pos.x = new_x;
-        self.cursor.pos.y = new_buffer_y;
-        self.debug_assert_invariants();
-    }
-
-    /// Erase from cursor to end of display (ED 0)
-    pub fn erase_to_end_of_display(&mut self) {
-        let cursor_y = self.cursor.pos.y;
-        let cursor_x = self.cursor.pos.x;
-
-        // PTY always operates at live bottom (scroll_offset = 0)
-        let visible_start = self.visible_window_start(0);
-        let visible_end = visible_start + self.height;
-
-        // Sweep image cells that are about to be erased (clears non-Kitty
-        // images buffer-wide and decrements image_cell_count for them).
-        self.collect_and_clear_image_ids_in_rows(cursor_y, visible_end, Some(cursor_x), None);
-
-        // Count any remaining image cells (e.g. Kitty images) that survived the
-        // sweep but will be destroyed by the row-level clear operations below.
-        let mut remaining_images = 0usize;
-        if self.image_cell_count > 0 {
-            if cursor_y < self.rows.len() {
-                remaining_images +=
-                    self.rows[cursor_y].count_image_cells_in_range(cursor_x, self.width);
-            }
-            for i in (cursor_y + 1)..visible_end.min(self.rows.len()) {
-                remaining_images += self.rows[i].count_image_cells();
-            }
-        }
-
-        // Clear from cursor to end of current row
-        if cursor_y < self.rows.len() {
-            let tag = self.current_tag.clone();
-            self.rows[cursor_y].clear_from(cursor_x, &tag);
-        }
-
-        // Clear all rows below cursor in the visible window
-        for i in (cursor_y + 1)..visible_end.min(self.rows.len()) {
-            let tag = self.current_tag.clone();
-            self.rows[i].clear_with_tag(&tag);
-        }
-
-        self.image_cell_count -= remaining_images;
-        self.debug_assert_invariants();
-    }
-
-    /// Erase from beginning of display to cursor (ED 1)
-    pub fn erase_to_beginning_of_display(&mut self) {
-        let cursor_y = self.cursor.pos.y;
-        let cursor_x = self.cursor.pos.x;
-
-        // PTY always operates at live bottom (scroll_offset = 0)
-        let visible_start = self.visible_window_start(0);
-
-        // Sweep image cells that are about to be erased (clears non-Kitty
-        // images buffer-wide and decrements image_cell_count for them).
-        // First row being erased starts at col 0; the cursor row is erased
-        // from col 0..=cursor_x.  For simplicity, sweep the entire row range.
-        self.collect_and_clear_image_ids_in_rows(visible_start, cursor_y + 1, None, None);
-
-        // Count any remaining image cells (e.g. Kitty images) that survived the
-        // sweep but will be destroyed by the row-level clear operations below.
-        let mut remaining_images = 0usize;
-        if self.image_cell_count > 0 {
-            for i in visible_start..cursor_y.min(self.rows.len()) {
-                remaining_images += self.rows[i].count_image_cells();
-            }
-            if cursor_y < self.rows.len() {
-                remaining_images += self.rows[cursor_y].count_image_cells_in_range(0, cursor_x + 1);
-            }
-        }
-
-        // Clear all rows above cursor in the visible window
-        for i in visible_start..cursor_y.min(self.rows.len()) {
-            let tag = self.current_tag.clone();
-            self.rows[i].clear_with_tag(&tag);
-        }
-
-        // Clear from beginning of current row to cursor
-        if cursor_y < self.rows.len() {
-            let tag = self.current_tag.clone();
-            self.rows[cursor_y].clear_to(cursor_x + 1, &tag);
-        }
-
-        self.image_cell_count -= remaining_images;
-        self.debug_assert_invariants();
-    }
-
-    /// Erase entire display (ED 2)
-    pub fn erase_display(&mut self) {
-        // PTY always operates at live bottom (scroll_offset = 0)
-        let visible_start = self.visible_window_start(0);
-        let visible_end = visible_start + self.height;
-
-        // Sweep image cells that are about to be erased (clears non-Kitty
-        // images buffer-wide and decrements image_cell_count for them).
-        self.collect_and_clear_image_ids_in_rows(visible_start, visible_end, None, None);
-
-        // Count any remaining image cells (e.g. Kitty images) that survived
-        // the sweep but will be destroyed by the row-level clear below.
-        let mut remaining_images = 0usize;
-        if self.image_cell_count > 0 {
-            for i in visible_start..visible_end.min(self.rows.len()) {
-                remaining_images += self.rows[i].count_image_cells();
-            }
-        }
-
-        for i in visible_start..visible_end.min(self.rows.len()) {
-            let tag = self.current_tag.clone();
-            self.rows[i].clear_with_tag(&tag);
-        }
-
-        self.image_cell_count -= remaining_images;
-        self.debug_assert_invariants();
-    }
-
-    /// Erase scrollback (ED 3)
-    pub fn erase_scrollback(&mut self) {
-        if self.kind == BufferType::Alternate {
-            // Alternate buffer has no scrollback
-            return;
-        }
-
-        let visible_start = self.visible_window_start(0);
-
-        // Remove all scrollback rows (everything before visible window)
-        if visible_start > 0 {
-            // Account for image cells in the drained scrollback rows.
-            if self.image_cell_count > 0 {
-                let drained_images: usize = self.rows[..visible_start]
-                    .iter()
-                    .map(Row::count_image_cells)
-                    .sum();
-                self.image_cell_count -= drained_images;
-            }
-            self.rows.drain(0..visible_start);
-            self.row_cache.drain(0..visible_start);
-            self.adjust_prompt_rows(visible_start);
-
-            // Adjust cursor
-            if self.cursor.pos.y >= visible_start {
-                self.cursor.pos.y -= visible_start;
-            } else {
-                self.cursor.pos.y = 0;
-            }
-
-            // scroll_offset is owned by the caller (ViewState); erase_scrollback
-            // removes all rows before the visible window so the caller should
-            // reset their scroll_offset to 0 after calling this method.
-        }
-
-        self.debug_assert_invariants();
-    }
-
-    /// Erase from cursor to end of line (EL 0)
-    pub fn erase_line_to_end(&mut self) {
-        let cursor_y = self.cursor.pos.y;
-        let cursor_x = self.cursor.pos.x;
-
-        if cursor_y < self.rows.len() {
-            // Sweep image cells that are about to be erased (non-Kitty images).
-            self.collect_and_clear_image_ids_in_rows(cursor_y, cursor_y + 1, Some(cursor_x), None);
-
-            // Count remaining image cells (e.g. Kitty) in the range being cleared.
-            let remaining = if self.image_cell_count > 0 {
-                self.rows[cursor_y].count_image_cells_in_range(cursor_x, self.width)
-            } else {
-                0
-            };
-
-            let tag = self.current_tag.clone();
-            self.rows[cursor_y].clear_from(cursor_x, &tag);
-            self.image_cell_count -= remaining;
-        }
-
-        self.debug_assert_invariants();
-    }
-
-    /// Erase from beginning of line to cursor (EL 1)
-    pub fn erase_line_to_beginning(&mut self) {
-        let cursor_y = self.cursor.pos.y;
-        let cursor_x = self.cursor.pos.x;
-
-        if cursor_y < self.rows.len() {
-            // Sweep image cells that are about to be erased (non-Kitty images).
-            self.collect_and_clear_image_ids_in_rows(cursor_y, cursor_y + 1, None, None);
-
-            // Count remaining image cells (e.g. Kitty) in the range being cleared.
-            let remaining = if self.image_cell_count > 0 {
-                self.rows[cursor_y].count_image_cells_in_range(0, cursor_x + 1)
-            } else {
-                0
-            };
-
-            let tag = self.current_tag.clone();
-            self.rows[cursor_y].clear_to(cursor_x + 1, &tag);
-            self.image_cell_count -= remaining;
-        }
-
-        self.debug_assert_invariants();
-    }
-
-    /// Erase entire line (EL 2)
-    pub fn erase_line(&mut self) {
-        let cursor_y = self.cursor.pos.y;
-
-        if cursor_y < self.rows.len() {
-            // Sweep image cells that are about to be erased (non-Kitty images).
-            self.collect_and_clear_image_ids_in_rows(cursor_y, cursor_y + 1, None, None);
-
-            // Count remaining image cells (e.g. Kitty) in the row.
-            let remaining = if self.image_cell_count > 0 {
-                self.rows[cursor_y].count_image_cells()
-            } else {
-                0
-            };
-
-            let tag = self.current_tag.clone();
-            self.rows[cursor_y].clear_with_tag(&tag);
-            self.image_cell_count -= remaining;
-        }
-
-        self.debug_assert_invariants();
-    }
-
-    /// Convert the currently visible rows into a flat `(Vec<TChar>, Vec<FormatTag>)` pair
-    /// Convert visible rows (with the given `scroll_offset`) into flat
-    /// `(Vec<TChar>, Vec<FormatTag>)` suitable for the GUI renderer.
-    ///
-    /// Pass `scroll_offset = 0` when calling from the PTY thread (which always
-    /// operates at the live bottom).
-    ///
-    /// Takes `&mut self` because it updates the per-row cache and clears dirty
-    /// flags on rows that are freshly flattened.
-    #[must_use]
-    pub fn visible_as_tchars_and_tags(
-        &mut self,
-        scroll_offset: usize,
-    ) -> (Vec<TChar>, Vec<FormatTag>, Vec<usize>, Vec<usize>) {
-        let visible_start = self.visible_window_start(scroll_offset);
-        let visible_end = (visible_start + self.height).min(self.rows.len());
-        Self::rows_as_tchars_and_tags_cached(
-            &mut self.rows[visible_start..visible_end],
-            &mut self.row_cache[visible_start..visible_end],
-        )
-    }
-
-    /// Flatten all scrollback rows (everything before the visible window) into
-    /// a linear `(Vec<TChar>, Vec<FormatTag>, Vec<usize>, Vec<usize>)` tuple
-    /// using the same algorithm as [`Self::visible_as_tchars_and_tags`].
-    ///
-    /// Returns `(vec![], vec![], vec![], vec![])` for the alternate screen
-    /// buffer, which never accumulates scrollback.
-    ///
-    /// Pass `scroll_offset = 0` when calling from the PTY thread.
-    pub fn scrollback_as_tchars_and_tags(
-        &mut self,
-        scroll_offset: usize,
-    ) -> (Vec<TChar>, Vec<FormatTag>, Vec<usize>, Vec<usize>) {
-        // Alternate buffer has no scrollback.
-        if self.kind == BufferType::Alternate {
-            return (vec![], vec![], vec![], vec![]);
-        }
-
-        let visible_start = self.visible_window_start(scroll_offset);
-
-        if visible_start == 0 {
-            // No scrollback rows exist yet.
-            return (vec![], vec![], vec![], vec![]);
-        }
-
-        Self::rows_as_tchars_and_tags_cached(
-            &mut self.rows[..visible_start],
-            &mut self.row_cache[..visible_start],
-        )
-    }
-
-    /// Shared helper: flatten a slice of [`Row`]s into `(Vec<TChar>,
-    /// Vec<FormatTag>, Vec<usize>)`, using a per-row cache to skip rows that
-    /// have not changed since the last snapshot.
-    ///
-    /// For each row:
-    /// - If `row.dirty` or the cache entry is `None`, flatten the row, populate
-    ///   the cache entry, and call `row.mark_clean()`.
-    /// - Otherwise reuse the cached per-row `(chars, tags)` directly.
-    ///
-    /// Per-row tag offsets are stored relative to each row's own character
-    /// slice (starting at 0).  The merge step below re-computes global offsets
-    /// each time, so the cache never stores stale absolute positions.
-    ///
-    /// The returned tuple contains:
-    /// - `Vec<TChar>` — flat character data
-    /// - `Vec<FormatTag>` — merged format tags with global offsets
-    /// - `Vec<usize>` — row offsets (`row_offsets[r]` is the flat index where
-    ///   row `r` begins)
-    /// - `Vec<usize>` — URL tag indices (indices into the tags vec where
-    ///   `url.is_some()`)
-    fn rows_as_tchars_and_tags_cached(
-        rows: &mut [Row],
-        cache: &mut [Option<(Vec<TChar>, Vec<FormatTag>)>],
-    ) -> (Vec<TChar>, Vec<FormatTag>, Vec<usize>, Vec<usize>) {
-        // ── Step 1: ensure every row has an up-to-date cache entry ──────────
-        for (row, entry) in rows.iter_mut().zip(cache.iter_mut()) {
-            if row.dirty || entry.is_none() {
-                *entry = Some(Self::flatten_row(row));
-                row.mark_clean();
-            }
-        }
-
-        // ── Step 2: merge per-row results into the global flat vectors ───────
-        // Per-row tags have offsets relative to the start of that row's chars.
-        // We accumulate a running `global_offset` and re-base each tag.
-        let row_count = rows.len();
-        let mut chars: Vec<TChar> = Vec::new();
-        let mut tags: Vec<FormatTag> = Vec::new();
-        let mut row_offsets: Vec<usize> = Vec::with_capacity(row_count);
-
-        for (row_idx, entry) in cache.iter().enumerate() {
-            // Step 1 populated every entry unconditionally, so `None` cannot
-            // occur here.  We use `if let` to satisfy the no-unwrap/expect rule;
-            // the `else` branch is unreachable in practice.
-            if let Some((row_chars, row_tags)) = entry.as_ref() {
-                let global_offset = chars.len();
-
-                // Record the flat index where this row begins.
-                row_offsets.push(global_offset);
-
-                // Append this row's characters, adjusting tag offsets.
-                for row_tag in row_tags {
-                    let rebased = FormatTag {
-                        start: global_offset + row_tag.start,
-                        end: global_offset + row_tag.end,
-                        colors: row_tag.colors,
-                        font_weight: row_tag.font_weight,
-                        font_decorations: row_tag.font_decorations,
-                        url: row_tag.url.clone(),
-                        blink: row_tag.blink,
-                    };
-
-                    // Merge with the previous tag when format is identical and
-                    // the ranges are contiguous (same logic as the original helper).
-                    if let Some(last) = tags.last_mut() {
-                        if last.end == rebased.start && tags_same_format(last, &rebased) {
-                            last.end = rebased.end;
-                        } else {
-                            tags.push(rebased);
-                        }
-                    } else {
-                        tags.push(rebased);
-                    }
-                }
-
-                chars.extend_from_slice(row_chars);
-            }
-
-            // Append a NewLine separator after every row except the last.
-            let is_last_row = row_idx + 1 == row_count;
-            if !is_last_row {
-                let byte_pos = chars.len();
-                chars.push(TChar::NewLine);
-                if let Some(last) = tags.last_mut() {
-                    if last.end == byte_pos {
-                        last.end += 1;
-                    } else {
-                        tags.push(FormatTag {
-                            start: byte_pos,
-                            end: byte_pos + 1,
-                            ..FormatTag::default()
-                        });
-                    }
-                } else {
-                    tags.push(FormatTag {
-                        start: byte_pos,
-                        end: byte_pos + 1,
-                        ..FormatTag::default()
-                    });
-                }
-            }
-        }
-
-        // Guarantee at least one tag covering the full range.
-        if tags.is_empty() {
-            tags.push(FormatTag {
-                start: 0,
-                end: if chars.is_empty() {
-                    usize::MAX
-                } else {
-                    chars.len()
-                },
-                ..FormatTag::default()
-            });
-        } else if let Some(last) = tags.last_mut() {
-            last.end = chars.len();
-        }
-
-        let url_tag_indices = Self::collect_url_tag_indices(&tags);
-        (chars, tags, row_offsets, url_tag_indices)
-    }
-
-    /// Collect the indices of tags in `tags` that carry a URL.
-    ///
-    /// This is a cheap post-pass over the already-built tag vector — typically
-    /// O(tags) where tags is small.  The result enables the GUI to iterate
-    /// only URL-bearing tags instead of scanning all tags during hover
-    /// detection.
-    fn collect_url_tag_indices(tags: &[FormatTag]) -> Vec<usize> {
-        tags.iter()
-            .enumerate()
-            .filter_map(|(i, tag)| tag.url.as_ref().map(|_| i))
-            .collect()
-    }
-
-    /// Flatten a single [`Row`] into a `(Vec<TChar>, Vec<FormatTag>)` pair.
-    ///
-    /// Tag offsets are **row-relative** (start at 0 for the first character in
-    /// this row).  The caller is responsible for re-basing them into global
-    /// offsets when merging multiple rows.
-    fn flatten_row(row: &Row) -> (Vec<TChar>, Vec<FormatTag>) {
-        let mut chars: Vec<TChar> = Vec::new();
-        let mut tags: Vec<FormatTag> = Vec::new();
-
-        for cell in row.get_characters() {
-            // Skip wide-glyph continuation cells.
-            if cell.is_continuation() {
-                continue;
-            }
-
-            let byte_pos = chars.len();
-            chars.push(*cell.tchar());
-
-            let cell_tag = cell.tag();
-            if let Some(last) = tags.last_mut() {
-                if last.end == byte_pos && tags_same_format(last, cell_tag) {
-                    last.end += 1;
-                } else {
-                    tags.push(FormatTag {
-                        start: byte_pos,
-                        end: byte_pos + 1,
-                        colors: cell_tag.colors,
-                        font_weight: cell_tag.font_weight,
-                        font_decorations: cell_tag.font_decorations,
-                        url: cell_tag.url.clone(),
-                        blink: cell_tag.blink,
-                    });
-                }
-            } else {
-                tags.push(FormatTag {
-                    start: byte_pos,
-                    end: byte_pos + 1,
-                    colors: cell_tag.colors,
-                    font_weight: cell_tag.font_weight,
-                    font_decorations: cell_tag.font_decorations,
-                    url: cell_tag.url.clone(),
-                    blink: cell_tag.blink,
-                });
-            }
-        }
-
-        // Guarantee at least one tag even for an empty row.
-        if tags.is_empty() {
-            tags.push(FormatTag {
-                start: 0,
-                end: 0,
-                ..FormatTag::default()
-            });
-        }
-
-        (chars, tags)
-    }
-
-    /// Return `true` when the alternate screen is currently active.
-    #[must_use]
-    pub const fn is_alternate_screen(&self) -> bool {
-        matches!(self.kind, BufferType::Alternate)
-    }
-
-    /// Return `true` when a cursor has been saved via DECSC (ESC 7 / `\x1b[?1048h`).
-    #[must_use]
-    pub const fn has_saved_cursor(&self) -> bool {
-        self.saved_cursor.is_some()
-    }
-
-    /// Return the terminal width (columns).
-    #[must_use]
-    pub const fn terminal_width(&self) -> usize {
-        self.width
-    }
-
-    /// Return the terminal height (rows).
-    #[must_use]
-    pub const fn terminal_height(&self) -> usize {
-        self.height
-    }
-
-    /// Extract the text content of a selection range from the buffer.
-    ///
-    /// Coordinates are buffer-absolute row indices (0 = first row in the full
-    /// buffer including scrollback). Columns are 0-indexed cell positions.
-    /// The range is inclusive on both ends: `[start_row, start_col]` through
-    /// `[end_row, end_col]`.
-    ///
-    /// Trailing whitespace on each row is trimmed (standard terminal behaviour).
-    /// Rows are separated by `'\n'`.
-    #[must_use]
-    pub fn extract_text(
-        &self,
-        start_row: usize,
-        start_col: usize,
-        end_row: usize,
-        end_col: usize,
-    ) -> String {
-        use std::fmt::Write as _;
-
-        if start_row >= self.rows.len() {
-            return String::new();
-        }
-        let end_row = end_row.min(self.rows.len().saturating_sub(1));
-
-        let mut result = String::new();
-
-        for row_idx in start_row..=end_row {
-            let row = &self.rows[row_idx];
-            let cells = row.get_characters();
-
-            let col_begin = if row_idx == start_row { start_col } else { 0 };
-            let col_end = if row_idx == end_row {
-                end_col
-            } else {
-                cells.len().saturating_sub(1)
-            };
-
-            let mut row_text = String::new();
-            for col in col_begin..=col_end {
-                if col >= cells.len() {
-                    break;
-                }
-                let cell = &cells[col];
-                if cell.is_continuation() {
-                    continue;
-                }
-                let tc = cell.tchar();
-                if matches!(tc, TChar::NewLine) {
-                    break;
-                }
-                write!(&mut row_text, "{tc}").unwrap_or_default();
-            }
-
-            let trimmed = row_text.trim_end();
-            result.push_str(trimmed);
-
-            if row_idx < end_row {
-                result.push('\n');
-            }
-        }
-
-        result
-    }
-
-    /// Extract a rectangular block of text from the buffer.
-    ///
-    /// Every row from `start_row` to `end_row` (inclusive) is sampled between
-    /// the same `col_min`..=`col_max` column range, where
-    /// `col_min = start_col.min(end_col)` and `col_max = start_col.max(end_col)`.
-    /// Rows are joined with `\n`.  Trailing whitespace is trimmed per row.
-    ///
-    /// This is the copy behaviour for Alt+drag (block/rectangular) selections.
-    #[must_use]
-    pub fn extract_block_text(
-        &self,
-        start_row: usize,
-        start_col: usize,
-        end_row: usize,
-        end_col: usize,
-    ) -> String {
-        use std::fmt::Write as _;
-
-        if start_row >= self.rows.len() {
-            return String::new();
-        }
-        let end_row = end_row.min(self.rows.len().saturating_sub(1));
-        let col_min = start_col.min(end_col);
-        let col_max = start_col.max(end_col);
-
-        let mut result = String::new();
-
-        for row_idx in start_row..=end_row {
-            let row = &self.rows[row_idx];
-            let cells = row.get_characters();
-
-            let mut row_text = String::new();
-            for col in col_min..=col_max {
-                if col >= cells.len() {
-                    break;
-                }
-                let cell = &cells[col];
-                if cell.is_continuation() {
-                    continue;
-                }
-                let tc = cell.tchar();
-                if matches!(tc, TChar::NewLine) {
-                    break;
-                }
-                write!(&mut row_text, "{tc}").unwrap_or_default();
-            }
-
-            let trimmed = row_text.trim_end();
-            result.push_str(trimmed);
-
-            if row_idx < end_row {
-                result.push('\n');
-            }
-        }
-
-        result
-    }
-
-    /// Set the current format tag for subsequent text insertions
-    pub fn set_format(&mut self, tag: FormatTag) {
-        self.current_tag = tag;
-    }
-
-    /// Get the current format tag
-    #[must_use]
-    pub const fn get_format(&self) -> &FormatTag {
-        &self.current_tag
-    }
-
-    /// Set whether Line Feed Mode (LNM) is enabled.
-    ///
-    /// `Lnm::NewLine`: LF behaves like CRLF (cursor moves to column 0 on line feed).
-    /// `Lnm::LineFeed` (default): LF only advances the row; column is unchanged.
-    pub const fn set_lnm(&mut self, mode: Lnm) {
-        self.lnm_enabled = mode;
-    }
-
-    /// Return whether Line Feed Mode is currently enabled.
-    #[must_use]
-    pub const fn is_lnm_enabled(&self) -> Lnm {
-        self.lnm_enabled
-    }
-
-    /// Set whether soft-wrapping is enabled (DECAWM).
-    ///
-    /// `Decawm::AutoWrap` (default): text wraps at the terminal width onto the next row.
-    /// `Decawm::NoAutoWrap`: text is clamped to the last column; any characters that would
-    /// overflow the current row are discarded.
-    pub const fn set_wrap(&mut self, mode: Decawm) {
-        self.wrap_enabled = mode;
-    }
-
-    /// Return whether soft-wrapping is currently enabled.
-    #[must_use]
-    pub const fn is_wrap_enabled(&self) -> Decawm {
-        self.wrap_enabled
-    }
-
-    /// Enable or disable DECOM (Origin Mode).
-    ///
-    /// When origin mode is enabled, cursor positioning (CUP, HVP, and similar
-    /// commands) is relative to the scroll region rather than the full screen.
-    /// Cursor movement is also constrained to the scroll region boundaries.
-    ///
-    /// Per DEC spec: enabling or disabling DECOM homes the cursor.
-    pub fn set_decom(&mut self, mode: Decom) {
-        self.decom_enabled = mode;
-        // DEC spec: changing DECOM homes the cursor to position (1,1) in the
-        // current addressing mode.  With `set_cursor_pos` already DECOM-aware,
-        // passing (0, 0) does the right thing for both modes.
-        self.set_cursor_pos(Some(0), Some(0));
-    }
-
-    /// Return whether DECOM (Origin Mode) is currently enabled.
-    #[must_use]
-    pub const fn is_decom_enabled(&self) -> Decom {
-        self.decom_enabled
-    }
-
-    /// Switch the buffer to a different column width (DECCOLM).
-    ///
-    /// This performs all the side effects mandated by the DEC spec:
-    /// - Reflows the buffer to the new width
-    /// - Clears the screen
-    /// - Resets the scroll region to full screen
-    /// - Resets DECOM (Origin Mode)
-    /// - Homes the cursor to (0, 0)
-    pub fn set_column_mode(&mut self, columns: usize) {
-        // Reflow to the new width (height stays the same).
-        // set_size returns the new scroll offset; DECCOLM always operates at
-        // live bottom, so pass 0 and discard the result.
-        let _ = self.set_size(columns, self.height, 0);
-
-        // Clear the visible screen.
-        self.erase_display();
-
-        // Reset scroll region to full screen.
-        self.scroll_region_top = 0;
-        self.scroll_region_bottom = self.height.saturating_sub(1);
-
-        // Reset DECOM without the cursor-home side effect (we home below).
-        self.decom_enabled = Decom::NormalCursor;
-
-        // Home cursor.
-        self.set_cursor_pos(Some(0), Some(0));
-    }
-
-    /// Switch to the alternate screen buffer.
-    ///
-    /// Enter the alternate screen buffer.
-    ///
-    /// Saves the primary buffer state (rows, cursor, scroll region, image store)
-    /// so it can be restored by [`Self::leave_alternate`]. Tab stops are NOT saved —
-    /// they are shared between primary and alternate screens (matching xterm).
-    ///
-    /// The caller must pass the current `scroll_offset` from `ViewState` so it can
-    /// be saved and restored later.  The alternate screen always starts at offset 0;
-    /// the caller should set `ViewState::scroll_offset = 0` after this call.
-    pub fn enter_alternate(&mut self, scroll_offset: usize) {
-        // If we're already in the alternate buffer, do nothing.
-        if self.kind == BufferType::Alternate {
-            return;
-        }
-
-        // Save primary state (rows + cursor + scroll_offset + cache).
-        let saved = SavedPrimaryState {
-            rows: self.rows.clone(),
-            row_cache: self.row_cache.clone(),
-            cursor: self.cursor.clone(),
-            scroll_offset,
-            height: self.height,
-            scroll_region_top: self.scroll_region_top,
-            scroll_region_bottom: self.scroll_region_bottom,
-            scroll_region_left: self.scroll_region_left,
-            scroll_region_right: self.scroll_region_right,
-            saved_cursor: self.saved_cursor.clone(),
-            image_store: self.image_store.clone(),
-            image_cell_count: self.image_cell_count,
-        };
-        self.saved_primary = Some(saved);
-
-        // Switch to alternate buffer.
-        self.kind = BufferType::Alternate;
-
-        // Fresh screen: exactly `height` empty rows, all dirty (None cache entries).
-        self.rows = vec![Row::new(self.width); self.height];
-        self.row_cache = vec![None; self.height];
-
-        // Alternate screen has no images.
-        self.image_store.clear();
-        self.image_cell_count = 0;
-
-        // Reset cursor for the alternate screen.
-        self.cursor = CursorState::default();
-
-        // Alternate screen starts with a full-screen scroll region.
-        self.reset_scroll_region_to_full();
-
-        self.debug_assert_invariants();
-    }
-
-    /// Leave the alternate screen and restore the primary buffer.
-    ///
-    /// Restores rows, cursor, scroll region, and image store from the saved
-    /// primary state. Tab stops are NOT restored — they are shared between
-    /// primary and alternate screens, so any changes made in the alternate
-    /// screen persist (matching xterm behavior).
-    ///
-    /// Returns the `scroll_offset` that was saved when `enter_alternate` was
-    /// called.  The caller should store this back into `ViewState::scroll_offset`.
-    /// Returns `0` if there was no saved primary state.
-    pub fn leave_alternate(&mut self) -> usize {
-        // Already in the primary buffer — nothing to do.
-        if self.kind == BufferType::Primary {
-            return 0;
-        }
-
-        self.kind = BufferType::Primary;
-
-        if let Some(saved) = self.saved_primary.take() {
-            // Restore saved primary state.
-            let restored_offset = saved.scroll_offset;
-            self.rows = saved.rows;
-            self.row_cache = saved.row_cache;
-            self.cursor = saved.cursor;
-            self.scroll_region_top = saved.scroll_region_top;
-            self.scroll_region_bottom = saved.scroll_region_bottom;
-            self.scroll_region_left = saved.scroll_region_left;
-            self.scroll_region_right = saved.scroll_region_right;
-            self.saved_cursor = saved.saved_cursor;
-            self.image_store = saved.image_store;
-            self.image_cell_count = saved.image_cell_count;
-
-            self.debug_assert_invariants();
-            restored_offset
-        } else {
-            self.debug_assert_invariants();
-            0
-        }
-    }
-
-    // ========================================================================
-    // Inline image support
-    // ========================================================================
-
-    /// Access the image store (read-only).
-    #[must_use]
-    pub const fn image_store(&self) -> &ImageStore {
-        &self.image_store
-    }
-
-    /// Access the image store (mutable).
-    pub const fn image_store_mut(&mut self) -> &mut ImageStore {
-        &mut self.image_store
-    }
-
-    /// Clear all image placements from every cell in the buffer.
-    pub fn clear_all_image_placements(&mut self) {
-        let mut cleared = 0usize;
-        for (i, row) in self.rows.iter_mut().enumerate() {
-            let mut changed = false;
-            for cell in row.cells_mut() {
-                if cell.has_image() {
-                    cell.clear_image();
-                    cleared += 1;
-                    changed = true;
-                }
-            }
-            if changed {
-                row.dirty = true;
-                if i < self.row_cache.len() {
-                    self.row_cache[i] = None;
-                }
-            }
-        }
-        self.image_cell_count -= cleared;
-    }
-
-    /// Clear all image placements for a specific image ID from every cell.
-    pub fn clear_image_placements_by_id(&mut self, image_id: u64) {
-        let mut cleared = 0usize;
-        for (i, row) in self.rows.iter_mut().enumerate() {
-            let mut changed = false;
-            for cell in row.cells_mut() {
-                if cell
-                    .image_placement()
-                    .is_some_and(|p| p.image_id == image_id)
-                {
-                    cell.clear_image();
-                    cleared += 1;
-                    changed = true;
-                }
-            }
-            if changed {
-                row.dirty = true;
-                if i < self.row_cache.len() {
-                    self.row_cache[i] = None;
-                }
-            }
-        }
-        self.image_cell_count -= cleared;
-    }
-
-    /// Clear image placements at the current cursor row and all rows after.
-    ///
-    /// Used by Kitty `d=c` (at cursor) and `d=C` (at cursor and after) delete
-    /// targets. Clears every image cell from the cursor row to the end of
-    /// the buffer.
-    pub fn clear_image_placements_at_cursor_and_after(&mut self) {
-        let start_row = self.cursor.pos.y;
-        let mut cleared = 0usize;
-        for i in start_row..self.rows.len() {
-            let row = &mut self.rows[i];
-            let mut changed = false;
-            for cell in row.cells_mut() {
-                if cell.has_image() {
-                    cell.clear_image();
-                    cleared += 1;
-                    changed = true;
-                }
-            }
-            if changed {
-                row.dirty = true;
-                if i < self.row_cache.len() {
-                    self.row_cache[i] = None;
-                }
-            }
-        }
-        self.image_cell_count -= cleared;
-    }
-
-    /// Clear image placements at the current cursor position only (single row).
-    pub fn clear_image_placements_at_cursor(&mut self) {
-        let row_idx = self.cursor.pos.y;
-        if row_idx >= self.rows.len() {
-            return;
-        }
-        let row = &mut self.rows[row_idx];
-        let mut cleared = 0usize;
-        for cell in row.cells_mut() {
-            if cell.has_image() {
-                cell.clear_image();
-                cleared += 1;
-            }
-        }
-        if cleared > 0 {
-            row.dirty = true;
-            if row_idx < self.row_cache.len() {
-                self.row_cache[row_idx] = None;
-            }
-            self.image_cell_count -= cleared;
-        }
-    }
-
-    /// Returns `true` if any cell in the buffer has an image placement.
-    ///
-    /// O(1) — backed by the `image_cell_count` counter.
-    #[must_use]
-    pub const fn has_any_image_cell(&self) -> bool {
-        self.image_cell_count > 0
-    }
-
-    /// Clear all image placements whose Kitty image number matches `number`.
-    pub fn clear_image_placements_by_number(&mut self, number: u32) {
-        let mut cleared = 0usize;
-        for (i, row) in self.rows.iter_mut().enumerate() {
-            let mut changed = false;
-            for cell in row.cells_mut() {
-                if cell
-                    .image_placement()
-                    .is_some_and(|p| p.image_number == Some(number))
-                {
-                    cell.clear_image();
-                    cleared += 1;
-                    changed = true;
-                }
-            }
-            if changed {
-                row.dirty = true;
-                if i < self.row_cache.len() {
-                    self.row_cache[i] = None;
-                }
-            }
-        }
-        self.image_cell_count -= cleared;
-    }
-
-    /// Clear image placements at a specific cell position.
-    pub fn clear_image_placements_at_cell(&mut self, row: usize, col: usize) {
-        if row >= self.rows.len() {
-            return;
-        }
-        let id = {
-            let cells = self.rows[row].cells();
-            if col < cells.len() {
-                cells[col].image_placement().map(|p| p.image_id)
-            } else {
-                None
-            }
-        };
-        if let Some(id) = id {
-            self.clear_image_placements_by_id(id);
-        }
-    }
-
-    /// Clear image placements at a specific cell and all cells after it.
-    pub fn clear_image_placements_at_cell_and_after(&mut self, row: usize, col: usize) {
-        let mut ids_to_clear: Vec<u64> = Vec::new();
-        for r in row..self.rows.len() {
-            let start = if r == row { col } else { 0 };
-            let cells = self.rows[r].cells();
-            for cell in cells.iter().skip(start) {
-                if let Some(placement) = cell.image_placement() {
-                    let id = placement.image_id;
-                    if !ids_to_clear.contains(&id) {
-                        ids_to_clear.push(id);
-                    }
-                }
-            }
-        }
-        for id in ids_to_clear {
-            self.clear_image_placements_by_id(id);
-        }
-    }
-
-    /// Clear all image placements that intersect the given column.
-    pub fn clear_image_placements_in_column(&mut self, col: usize) {
-        let mut ids_to_clear: Vec<u64> = Vec::new();
-        for row in &self.rows {
-            let cells = row.cells();
-            if col < cells.len()
-                && let Some(placement) = cells[col].image_placement()
-            {
-                let id = placement.image_id;
-                if !ids_to_clear.contains(&id) {
-                    ids_to_clear.push(id);
-                }
-            }
-        }
-        for id in ids_to_clear {
-            self.clear_image_placements_by_id(id);
-        }
-    }
-
-    /// Clear all image placements that intersect the given row.
-    pub fn clear_image_placements_in_row(&mut self, row: usize) {
-        if row >= self.rows.len() {
-            return;
-        }
-        let mut ids_to_clear: Vec<u64> = Vec::new();
-        let cells = self.rows[row].cells();
-        for cell in cells {
-            if let Some(placement) = cell.image_placement() {
-                let id = placement.image_id;
-                if !ids_to_clear.contains(&id) {
-                    ids_to_clear.push(id);
-                }
-            }
-        }
-        for id in ids_to_clear {
-            self.clear_image_placements_by_id(id);
-        }
-    }
-
-    /// Clear all image placements with the given z-index.
-    pub fn clear_image_placements_by_z_index(&mut self, z: i32) {
-        let mut cleared = 0usize;
-        for (i, row) in self.rows.iter_mut().enumerate() {
-            let mut changed = false;
-            for cell in row.cells_mut() {
-                if cell.image_placement().is_some_and(|p| p.z_index == z) {
-                    cell.clear_image();
-                    cleared += 1;
-                    changed = true;
-                }
-            }
-            if changed {
-                row.dirty = true;
-                if i < self.row_cache.len() {
-                    self.row_cache[i] = None;
-                }
-            }
-        }
-        self.image_cell_count -= cleared;
-    }
-    /// would overwrite any image cells.  If so, clear **all** cells of each
-    /// affected image across the entire buffer.
-    ///
-    /// Only the cells in `[col .. col + text_len)` are inspected — images
-    /// outside this range on the same row are not affected.
-    ///
-    /// Images are treated as atomic: overwriting even a single cell of an
-    /// image invalidates the whole image.  This matches the behaviour of
-    /// other terminal emulators (`iTerm2`, `WezTerm`, Kitty).
-    fn clear_images_overwritten_by_text(&mut self, row_idx: usize, col: usize, text_len: usize) {
-        self.collect_and_clear_image_ids_in_rows(
-            row_idx,
-            row_idx + 1,
-            Some(col),
-            Some(col + text_len),
-        );
-    }
-
-    /// Scan rows in `[row_start..row_end)` for image cells and clear every
-    /// cell of each found image across the entire buffer.
-    ///
-    /// If `start_col` is `Some(col)`, only cells from `col..` on the first
-    /// row (or single row) are scanned.  If `end_col` is `Some(ec)`, only
-    /// cells up to (exclusive) `ec` are scanned on the first row.  If
-    /// `None`, scanning goes to the end of the row.
-    ///
-    /// For rows after the first (when `row_end - row_start > 1`), all cells
-    /// are always scanned.
-    ///
-    /// This is a no-op when no image cells are present.
-    fn collect_and_clear_image_ids_in_rows(
-        &mut self,
-        row_start: usize,
-        row_end: usize,
-        start_col: Option<usize>,
-        end_col: Option<usize>,
-    ) {
-        let end = row_end.min(self.rows.len());
-        if row_start >= end {
-            return;
-        }
-
-        // Collect unique image IDs from the targeted cells, skipping Kitty
-        // images (they are cleared only via explicit `a=d` commands).
-        let mut ids_to_clear: Vec<u64> = Vec::new();
-        for (idx, row) in self.rows[row_start..end].iter().enumerate() {
-            let cells = row.cells();
-            let skip = if idx == 0 {
-                start_col.unwrap_or(0).min(cells.len())
-            } else {
-                0
-            };
-            let limit = if idx == 0 {
-                end_col.unwrap_or(cells.len()).min(cells.len())
-            } else {
-                cells.len()
-            };
-            if skip >= limit {
-                continue;
-            }
-            for cell in &cells[skip..limit] {
-                if let Some(placement) = cell.image_placement() {
-                    // Kitty images are not cleared by text writes.
-                    if placement.protocol == ImageProtocol::Kitty {
-                        continue;
-                    }
-                    let id = placement.image_id;
-                    if !ids_to_clear.contains(&id) {
-                        ids_to_clear.push(id);
-                    }
-                }
-            }
-        }
-
-        // Clear all cells of each affected image buffer-wide.
-        for id in ids_to_clear {
-            self.clear_image_placements_by_id(id);
-        }
-    }
-
-    /// Place an inline image at the current cursor position.
-    ///
-    /// The image is stored in the central `ImageStore` and cells in the
-    /// rectangular region `[cursor_y .. cursor_y + display_rows) ×
-    /// [cursor_x .. cursor_x + display_cols)` are filled with
-    /// `ImagePlacement` references.
-    ///
-    /// After placement the cursor is moved to the row immediately below
-    /// the image (or the last visible row if the image extends to the
-    /// bottom), at column 0 — matching iTerm2 behaviour.
-    ///
-    /// If the image extends beyond the right edge of the terminal, it is
-    /// clipped to the terminal width (cells beyond the edge are not placed).
-    /// If the image extends below the visible area, new rows are created
-    /// (scrolling if necessary in the primary buffer).
-    pub fn place_image(
-        &mut self,
-        image: InlineImage,
-        scroll_offset: usize,
-        protocol: ImageProtocol,
-        image_number: Option<u32>,
-        placement_id: Option<u32>,
-        z_index: i32,
-    ) -> usize {
-        let image_id = image.id;
-        let display_cols = image.display_cols;
-        let display_rows = image.display_rows;
-
-        // Store the image centrally.
-        self.image_store.insert(image);
-
-        let start_col = self.cursor.pos.x;
-
-        // Clamp display_cols to not exceed terminal width.
-        let effective_cols = display_cols.min(self.width.saturating_sub(start_col));
-
-        // Before placing new image cells, clear any existing image cells in
-        // the column range from the cursor row downward.  This handles the
-        // common case where a new (possibly smaller) image replaces an old
-        // (possibly larger) one at the same position — without this, stale
-        // cells from the old image persist below the new one.
-        let clear_end_col = start_col + effective_cols;
-        for row_idx in self.cursor.pos.y..self.rows.len() {
-            let row = &mut self.rows[row_idx];
-            let mut changed = false;
-            for col in start_col..clear_end_col.min(row.max_width()) {
-                if let Some(cell) = row.cells_mut().get_mut(col)
-                    && cell.has_image()
-                {
-                    cell.clear_image();
-                    self.image_cell_count -= 1;
-                    changed = true;
-                }
-            }
-            if changed {
-                row.dirty = true;
-                if row_idx < self.row_cache.len() {
-                    self.row_cache[row_idx] = None;
-                }
-            }
-        }
-
-        let mut current_offset = scroll_offset;
-
-        // Place image cells row by row.
-        //
-        // We track `base_row` which starts at the cursor's current row and
-        // shifts downward as rows are created.  Unlike `scroll_up()`, we
-        // grow the buffer by pushing new rows and then let
-        // `enforce_scrollback_limit` trim excess from the top — this avoids
-        // the infinite-loop problem where `scroll_up()` keeps rows.len()
-        // constant.
-        let mut base_row = self.cursor.pos.y;
-
-        for img_row in 0..display_rows {
-            let target_row = base_row + img_row;
-
-            // Ensure the target row exists.
-            while target_row >= self.rows.len() {
-                self.push_row(RowOrigin::HardBreak, RowJoin::NewLogicalLine);
-            }
-
-            let row = &mut self.rows[target_row];
-            row.dirty = true;
-
-            // Invalidate the row cache for this row.
-            if target_row < self.row_cache.len() {
-                self.row_cache[target_row] = None;
-            }
-
-            let mut placed_count = 0usize;
-            for img_col in 0..effective_cols {
-                let col = start_col + img_col;
-                if col >= self.width {
-                    break;
-                }
-                let placement = ImagePlacement {
-                    image_id,
-                    col_in_image: img_col,
-                    row_in_image: img_row,
-                    protocol,
-                    image_number,
-                    placement_id,
-                    z_index,
-                };
-                row.set_image_cell(col, placement, self.current_tag.clone());
-                placed_count += 1;
-            }
-            self.image_cell_count += placed_count;
-        }
-
-        // Enforce scrollback limit — this may drain rows from the top.
-        if self.kind == BufferType::Primary {
-            let rows_before = self.rows.len();
-            current_offset = self.enforce_scrollback_limit(current_offset);
-            let drained = rows_before - self.rows.len();
-            // Adjust base_row for the drained rows.
-            base_row = base_row.saturating_sub(drained);
-        }
-
-        // Move cursor below the image, column 0 (iTerm2 behaviour).
-        let final_row = base_row + display_rows;
-        if final_row < self.rows.len() {
-            self.cursor.pos.y = final_row;
-        } else {
-            self.cursor.pos.y = self.rows.len().saturating_sub(1);
-        }
-        self.cursor.pos.x = 0;
-
-        self.debug_assert_invariants();
-        current_offset
-    }
 }
 
 // ============================================================================
@@ -3938,7 +463,7 @@ impl Buffer {
 
 /// Compare two `FormatTag` values by their visual formatting only, ignoring
 /// the `start` and `end` byte-position fields.
-fn tags_same_format(a: &FormatTag, b: &FormatTag) -> bool {
+pub(in crate::buffer) fn tags_same_format(a: &FormatTag, b: &FormatTag) -> bool {
     a.colors == b.colors
         && a.font_weight == b.font_weight
         && a.font_decorations == b.font_decorations
